+++ _pkgdown.yml
@@ -1,7 +1,6 @@
 url: https://easystats.github.io/datawizard/
 
 template:
-  bootstrap: 5
   package: easystatstemplate
 
 
