### 2022-08-10 196d3f1
Deploying to gh-pages from @ easystats/datawizard@40f25c5a49588624a3effb48ca155b302ea6e128 🚀
