### 2022-08-11 9790b84
turn on devel mode for pkgdown
