+++ DESCRIPTION
@@ -26,7 +26,7 @@ BugReports: https://github.com/easystats/datawizard/issues
 Depends:
     R (>= 3.5)
 Imports:
-    insight (>= 0.18.0),
+    insight (>= 0.18.2),
     stats,
     utils
 Suggests: 
@@ -52,7 +52,7 @@ Suggests:
     rmarkdown,
     rstanarm,
     see,
-    testthat (>= 3.0.0),
+    testthat (>= 3.1.0),
     tidyr,
     withr
 VignetteBuilder: 
