### 2022-08-12 29478f0
convert_na_to improvements when dealing with character
