+++ tests/testthat/test-convert_na_to.R
@@ -59,12 +59,12 @@ test_that("convert_na_to - character: works", {
 
 test_that("convert_na_to - character: arg 'replacement' can only be character", {
   expect_warning(
-    convert_na_to(c("a", "b", "c", NA), replacement = 1),
-    regexp = "`replacement` needs to be a character vector."
+    convert_na_to(c("a", "b", "c", NA), replacement = mtcars),
+    regexp = "`replacement` needs to be a character or numeric vector."
   )
   expect_warning(
     convert_na_to(c("a", "b", "c", NA), replacement = factor(8)),
-    regexp = "`replacement` needs to be a character vector."
+    regexp = "`replacement` needs to be a character or numeric vector."
   )
 })
 
@@ -77,7 +77,7 @@ test_that("convert_na_to - numeric: arg 'replacement' must be of length one", {
 
 test_that("convert_na_to - character: returns original vector if 'replacement' not good", {
   expect_equal(
-    convert_na_to(c("a", "b", "c", NA), replacement = 1, verbose = FALSE),
+    convert_na_to(c("a", "b", "c", NA), replacement = mtcars, verbose = FALSE),
     c("a", "b", "c", NA)
   )
   expect_equal(
