### 2022-08-12 33365a7
Updated news and description with changes and credit
