### 2022-08-12 42f8dcb
boolean && and test adjustment for numeric
