### 2022-08-13 4dc4a60
`convert_na_to` improvements when dealing with character (#215)
* convert_na_to improvements when dealing with character

* Updated news and description with changes and credit

* Correct typo in Description

* boolean && and test adjustment for numeric

Co-authored-by: RemPsyc <RemPsyc@users.noreply.github.com>
