### 2022-08-16 96f1783
Fix tests for `{poorman}` update
