### 2022-08-17 34cf6bf
same as JOSS cite; style; README cleanup
