+++ R/convert_na_to.R
@@ -123,7 +123,8 @@ convert_na_to.character <- function(x, replacement = NULL, verbose = TRUE, ...) 
   if (is_empty_object(replacement) || !is.character(replacement) && !is.numeric(replacement)) {
     if (isTRUE(verbose)) {
       warning(insight::format_message(
-        "`replacement` needs to be a character or numeric vector."), call. = FALSE)
+        "`replacement` needs to be a character or numeric vector."
+      ), call. = FALSE)
     }
   } else if (length(replacement) > 1) {
     if (isTRUE(verbose)) {
