+++ cran-comments.md
@@ -1,11 +1,8 @@
-## Test environments
-* local R installation, R 4.2.0
-* ubuntu 16.04 (on github-actions), R 4.2.0
-* win-builder (devel)
-
 ## R CMD check results
 
 0 errors | 0 warnings | 0 note
+
+* Fixes failing tests due to `{poorman}` update.
 
 ## revdepcheck results
 
