### 2022-08-18 0601442
Fix tests for `{poorman}` update (#220)
* Fix tests for `{poorman}` update

* use namespace

* poorman update on CRAN

* Ready for CRAN?
