### 2022-08-18 e2330e7
Deploying to gh-pages from @ easystats/datawizard@0601442ec5874b9ebcf3a056fbec01cbcfaaf59e 🚀
