### 2022-08-18 #4659 @osorensen
👋 @nhtran93 @njtierney @pmyteh @gabett would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
