### 2022-08-18 856b76a
Use recommended workflow for template package
