+++ DESCRIPTION
@@ -59,8 +59,12 @@ Suggests:
     withr
 VignetteBuilder: 
     knitr
-Config/testthat/edition: 3
 Encoding: UTF-8
 Language: en-US
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.1
+Config/testthat/edition: 3
+Config/Needs/website:
+    rstudio/bslib
+    r-lib/pkgdown,
+    easystats/easystatstemplate
