+++ .github/workflows/pkgdown.yml
@@ -29,8 +29,6 @@ jobs:
         with:
           extra-packages: |
               local::.
-              r-lib/pkgdown
-              easystats/easystatstemplate
           needs: website
 
       - name: Build site
