+++ DESCRIPTION
@@ -65,6 +65,6 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.1
 Config/testthat/edition: 3
 Config/Needs/website:
-    rstudio/bslib
+    rstudio/bslib,
     r-lib/pkgdown,
     easystats/easystatstemplate
