### 2022-08-18 d0bd456
Use recommended workflow for template package (#222)
* Use recommended workflow for template package

* oops
