+++ .github/workflows/pkgdown.yml
@@ -20,6 +20,8 @@ jobs:
       - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-pandoc@v2
+        with:
+            pandoc-version: '2.19'
 
       - uses: r-lib/actions/setup-r@v2
         with:
