+++ pkgdown.yml
@@ -1,10 +1,10 @@
-pandoc: 2.14.2
+pandoc: '2.19'
 pkgdown: 2.0.6.9000
 pkgdown_sha: 956f076f661eb2de9413f64ea24a580cf444466d
 articles:
   standardize_data: standardize_data.html
   tidyverse_translation: tidyverse_translation.html
-last_built: 2022-08-18T17:57Z
+last_built: 2022-08-18T18:55Z
 urls:
   reference: https://easystats.github.io/datawizard/reference
   article: https://easystats.github.io/datawizard/articles
