### 2022-08-19 #4684 @editorialbot
[REVIEW]: datawizard: An R Package for Easy Data Preparation and Statistical Transformations
**Submitting author:** @IndrajeetPatil (<a href="http://orcid.org/0000-0003-1995-6531">Indrajeet Patil</a>)
**Repository:** https://github.com/easystats/datawizard
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.2
