+++ .github/workflows/pkgdown.yml
@@ -21,7 +21,7 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
         with:
-            pandoc-version: '2.19'
+            pandoc-version: '2.19.1'
 
       - uses: r-lib/actions/setup-r@v2
         with:
