### 2022-08-19 #4659 @njtierney
Just clarifying if you still need a reviewer?

On Thu, 18 Aug 2022 at 7:20 pm, Øystein Sørensen ***@***.***>
wrote:

> 👋 @nhtran93 <https://github.com/nhtran93> @njtierney
> <https://github.com/njtierney> @pmyteh <https://github.com/pmyteh> @gabett
> <https://github.com/gabett> would any of you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
