### 2022-08-20 cbcf7f2
Redo reference section for website (#223)
