### 2022-08-20 7b921cb
Remove aliases
Closes #221
