+++ revdep/cran.md
@@ -1,37 +1,7 @@
 ## revdepcheck results
 
-We checked 13 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
+We checked 14 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
 
- * We saw 7 new problems
+ * We saw 0 new problems
  * We failed to check 0 packages
-
-Issues with CRAN packages are summarised below.
-
-### New problems
-(This reports the first line of each new failure)
-
-* correlation
-  checking examples ... ERROR
-
-* effectsize
-  checking tests ...
-
-* parameters
-  checking examples ... ERROR
-  checking tests ...
-  checking dependencies in R code ... WARNING
-
-* report
-  checking examples ... ERROR
-  checking R code for possible problems ... NOTE
-
-* see
-  checking examples ... ERROR
-  checking dependencies in R code ... WARNING
-
-* sjPlot
-  checking examples ... ERROR
-
-* statsExpressions
-  checking tests ...
 
