+++ revdep/README.md
@@ -3,34 +3,22 @@
 |field    |value                                     |
 |:--------|:-----------------------------------------|
 |version  |R version 4.2.1 (2022-06-23)              |
-|os       |macOS Monterey 12.4                       |
+|os       |macOS Monterey 12.5                       |
 |system   |aarch64, darwin20                         |
 |ui       |RStudio                                   |
 |language |(EN)                                      |
 |collate  |en_US.UTF-8                               |
 |ctype    |en_US.UTF-8                               |
 |tz       |Europe/Berlin                             |
-|date     |2022-08-06                                |
-|rstudio  |2022.07.0+548 Spotted Wakerobin (desktop) |
+|date     |2022-08-20                                |
+|rstudio  |2022.07.1+554 Spotted Wakerobin (desktop) |
 |pandoc   |2.19 @ /usr/local/bin/ (via rmarkdown)    |
 
 # Dependencies
 
-|package    |old   |new      |Δ  |
-|:----------|:-----|:--------|:--|
-|datawizard |0.4.1 |0.4.1.10 |*  |
+|package    |old   |new   |Δ  |
+|:----------|:-----|:-----|:--|
+|datawizard |0.5.1 |0.5.1 |NA |
 
 # Revdeps
-
-## New problems (7)
-
-|package                                          |version |error  |warning |note   |
-|:------------------------------------------------|:-------|:------|:-------|:------|
-|[correlation](problems.md#correlation)           |0.8.1   |__+1__ |        |       |
-|[effectsize](problems.md#effectsize)             |0.7.0   |__+1__ |        |       |
-|[parameters](problems.md#parameters)             |0.18.1  |__+2__ |__+1__  |       |
-|[report](problems.md#report)                     |0.5.1   |__+1__ |        |__+1__ |
-|[see](problems.md#see)                           |0.7.1   |__+1__ |__+1__  |       |
-|[sjPlot](problems.md#sjplot)                     |2.8.10  |__+1__ |        |       |
-|[statsExpressions](problems.md#statsexpressions) |1.3.2   |__+1__ |        |       |
 
