+++ NAMESPACE
@@ -168,7 +168,6 @@ export(convert_to_na)
 export(data_addprefix)
 export(data_addsuffix)
 export(data_adjust)
-export(data_cut)
 export(data_extract)
 export(data_filter)
 export(data_find)
@@ -178,7 +177,6 @@ export(data_match)
 export(data_merge)
 export(data_partition)
 export(data_read)
-export(data_recode)
 export(data_relocate)
 export(data_remove)
 export(data_rename)
@@ -187,11 +185,8 @@ export(data_reorder)
 export(data_restoretype)
 export(data_rotate)
 export(data_select)
-export(data_shift)
 export(data_tabulate)
-export(data_to_factor)
 export(data_to_long)
-export(data_to_numeric)
 export(data_to_wide)
 export(data_transpose)
 export(data_ungroup)
