### 2022-08-20 5428ad6
Remove `coerce_to_numeric()`
Closes #206
