+++ R/data_restoretype.R
@@ -26,7 +26,7 @@ data_restoretype <- function(data, reference = NULL, ...) {
     # No reference data (regular fixing) ----------------
     if (is.null(reference)) {
       if (is.character(data[[col]])) {
-        data[[col]] <- coerce_to_numeric(data[[col]])
+        data[[col]] <- as.numeric(data[[col]])
       }
     } else {
       if (is.factor(reference[[col]]) && !is.factor(data[[col]])) {
@@ -35,7 +35,7 @@ data_restoretype <- function(data, reference = NULL, ...) {
       }
 
       if (is.numeric(reference[[col]]) && !is.numeric(data[[col]])) {
-        data[[col]] <- coerce_to_numeric(as.character(data[[col]]))
+        data[[col]] <- as.numeric(as.character(data[[col]]))
       }
 
       if (is.character(reference[[col]]) && !is.character(data[[col]])) {
