+++ man/coerce_to_numeric.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/to_numeric.R
-\name{coerce_to_numeric}
-\alias{coerce_to_numeric}
-\title{Convert to Numeric (if possible)}
-\usage{
-coerce_to_numeric(x)
-}
-\arguments{
-\item{x}{A vector to be converted.}
-}
-\value{
-Numeric vector (if possible)
-}
-\description{
-Tries to convert vector to numeric if possible (if no warnings or errors).
-Otherwise, leaves it as is.
-}
-\examples{
-coerce_to_numeric(c("1", "2"))
-coerce_to_numeric(c("1", "2", "A"))
-}
