+++ NAMESPACE
@@ -158,7 +158,6 @@ export(center)
 export(centre)
 export(change_code)
 export(change_scale)
-export(coerce_to_numeric)
 export(colnames_to_row)
 export(column_as_rownames)
 export(compact_character)
