### 2022-08-20 39f9b73
Deprecate  `format_text()`
Closes #224
