+++ NAMESPACE
@@ -232,6 +232,7 @@ export(smoothness)
 export(standardise)
 export(standardize)
 export(text_concatenate)
+export(text_format)
 export(text_fullstop)
 export(text_lastchar)
 export(text_paste)
