### 2022-08-21 3970d46
Remove aliases (#225)
Closes #221
