+++ revdep/problems.md
@@ -1,340 +1,1 @@
-# correlation
-
-<details>
-
-* Version: 0.8.1
-* GitHub: https://github.com/easystats/correlation
-* Source code: https://github.com/cran/correlation
-* Date/Publication: 2022-05-20 21:50:02 UTC
-* Number of recursive dependencies: 184
-
-Run `revdep_details(, "correlation")` for more info
-
-</details>
-
-## Newly broken
-
-*   checking examples ... ERROR
-    ```
-    Running examples in ‘correlation-Ex.R’ failed
-    The error most likely occurred in:
-    
-    > ### Name: visualisation_recipe.easycor_test
-    > ### Title: Visualisation Recipe for 'correlation' Objects
-    > ### Aliases: visualisation_recipe.easycor_test
-    > ###   visualisation_recipe.easycormatrix
-    > ###   visualisation_recipe.easycorrelation
-    > 
-    > ### ** Examples
-    ...
-    +     labs = list(title = "My Plot")
-    +   )
-    +   plot(layers) + theme_modern()
-    + }
-    Loading required package: see
-    Warning: Argument `colnames_to` is deprecated. Please use `names_to` instead.
-    Error in colnames(data)[sapply(data, select)] : 
-      invalid subscript type 'list'
-    Calls: visualisation_recipe ... visualisation_recipe.easycormatrix -> <Anonymous> -> .select_nse
-    Execution halted
-    ```
-
-# effectsize
-
-<details>
-
-* Version: 0.7.0
-* GitHub: https://github.com/easystats/effectsize
-* Source code: https://github.com/cran/effectsize
-* Date/Publication: 2022-05-26 13:20:02 UTC
-* Number of recursive dependencies: 234
-
-Run `revdep_details(, "effectsize")` for more info
-
-</details>
-
-## Newly broken
-
-*   checking tests ...
-    ```
-      Running ‘spelling.R’
-      Running ‘testthat.R’
-     ERROR
-    Running the tests in ‘tests/testthat.R’ failed.
-    Last 13 lines of output:
-       7.       └─effectsize (local) FUN(newX[, i], ...)
-       8.         └─datawizard::ranktransform(x, method = "average", ..., verbose = FALSE)
-      ── Error (test-rankES.R:76:5): kendalls_w ──────────────────────────────────────
-      Error in `UseMethod("ranktransform")`: no applicable method for 'ranktransform' applied to an object of class "character"
-      Backtrace:
-    ...
-          ▆
-       1. └─effectsize::kendalls_w(M1) at test-rankES.R:76:4
-       2.   └─effectsize:::.kendalls_w(data, verbose = verbose)
-       3.     └─base::apply(data, 1, .safe_ranktransform, verbose = verbose)
-       4.       └─effectsize (local) FUN(newX[, i], ...)
-       5.         └─datawizard::ranktransform(x, method = "average", ..., verbose = FALSE)
-      
-      [ FAIL 3 | WARN 3 | SKIP 11 | PASS 494 ]
-      Error: Test failures
-      Execution halted
-    ```
-
-# parameters
-
-<details>
-
-* Version: 0.18.1
-* GitHub: https://github.com/easystats/parameters
-* Source code: https://github.com/cran/parameters
-* Date/Publication: 2022-05-29 05:50:02 UTC
-* Number of recursive dependencies: 374
-
-Run `revdep_details(, "parameters")` for more info
-
-</details>
-
-## Newly broken
-
-*   checking examples ... ERROR
-    ```
-    Running examples in ‘parameters-Ex.R’ failed
-    The error most likely occurred in:
-    
-    > ### Name: cluster_analysis
-    > ### Title: Cluster Analysis
-    > ### Aliases: cluster_analysis
-    > 
-    > ### ** Examples
-    > 
-    > set.seed(33)
[…]
