### 2022-08-21 2d7e441
make `coerce_to_numeric()` internal
