### 2022-08-21 94d940f
Deprecate  `format_text()` (#227)
Closes #224
