+++ man/text_format.Rd
@@ -1,6 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/format_text.R
-\name{format_text}
+\name{text_format}
+\alias{text_format}
 \alias{format_text}
 \alias{text_fullstop}
 \alias{text_lastchar}
@@ -10,6 +11,15 @@
 \alias{text_wrap}
 \title{Convenient text formatting functionalities}
 \usage{
+text_format(
+  text,
+  sep = ", ",
+  last = " and ",
+  width = NULL,
+  enclose = NULL,
+  ...
+)
+
 format_text(
   text,
   sep = ", ",
