+++ R/data_find.R
@@ -130,7 +130,9 @@ find_columns <- function(data,
   if (!length(columns) || is.null(columns)) {
     columns <- NULL
     if (isTRUE(verbose)) {
-      warning(insight::format_message("No column names that matched the required search pattern were found."), call. = FALSE)
+      warning(insight::format_message(
+        "No column names that matched the required search pattern were found."
+      ), call. = FALSE)
     }
   }
 
