+++ R/data_relocate.R
@@ -49,7 +49,7 @@ data_relocate <- function(data,
                           ...) {
   # Sanitize
   if (!is.null(before) && !is.null(after)) {
-    stop("You must supply only one of `before` or `after`.")
+    stop("You must supply only one of `before` or `after`.", call. = FALSE)
   }
 
   # allow numeric values
@@ -59,7 +59,7 @@ data_relocate <- function(data,
     } else if (before >= 1 && before <= ncol(data)) {
       before <- names(data)[before]
     } else {
-      stop("No valid position defined in 'before'.", call. = FALSE)
+      stop("No valid position defined in `before`.", call. = FALSE)
     }
   }
 
@@ -70,7 +70,7 @@ data_relocate <- function(data,
     } else if (after >= 1 && after <= ncol(data)) {
       after <- names(data)[after]
     } else {
-      stop("No valid position defined in 'after'.", call. = FALSE)
+      stop("No valid position defined in `after`.", call. = FALSE)
     }
   }
 
@@ -90,14 +90,14 @@ data_relocate <- function(data,
   if (!is.null(before)) {
     before <- before[before %in% data_cols][1] # Take first that exists (if vector is supplied)
     if (length(before) != 1) {
-      stop("The column passed to 'before' wasn't found. Possibly mispelled.")
+      stop("The column passed to `before` wasn't found. Possibly mispelled.", call. = FALSE)
     }
     where <- min(match(before, data_cols))
     position <- c(setdiff(position, where), where)
   } else if (!is.null(after)) {
     after <- after[after %in% data_cols][1] # Take first that exists (if vector is supplied)
     if (length(after) != 1) {
-      stop("The column passed to 'after' wasn't found. Possibly mispelled.")
+      stop("The column passed to `after` wasn't found. Possibly mispelled.", call. = FALSE)
     }
     where <- max(match(after, data_cols))
     position <- c(where, setdiff(position, where))
