+++ .github/workflows/R-CMD-check.yaml
@@ -30,10 +30,10 @@ jobs:
           #- {os: windows-latest, r: '3.6.0'}
           #- {os: windows-latest, r: '3.5.0'}
 
-          - {os: ubuntu-18.04,   r: 'devel', http-user-agent: 'release'}
-          - {os: ubuntu-18.04,   r: 'release'}
-          - {os: ubuntu-18.04,   r: 'oldrel'}
-          #- {os: ubuntu-18.04,   r: '3.6.0'}
+          - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
+          - {os: ubuntu-latest,  r: 'release'}
+          - {os: ubuntu-latest,  r: 'oldrel'}
+          #- {os: ubuntu-latest,   r: '3.6.0'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
