### 2022-08-23 b939768
Deploying to gh-pages from @ easystats/datawizard@8c66bc244b4cd90711ee19a7ffeed9993eda2b34 🚀
