### 2022-08-23 3553b38
add regex argument to data_remove
