+++ R/data_remove.R
@@ -5,11 +5,23 @@
 #' head(data_remove(iris, "Sepal.Length"))
 #' head(data_remove(iris, starts_with("Sepal")))
 #' @export
-data_remove <- function(data, select, ignore_case = FALSE, verbose = FALSE, ...) {
+data_remove <- function(data,
+                        select = NULL,
+                        exclude = NULL,
+                        ignore_case = FALSE,
+                        regex = FALSE,
+                        verbose = FALSE,
+                        ...) {
   ## TODO set verbose = TRUE by default in a later update?
 
   # evaluate arguments
-  select <- .select_nse(select, data, exclude = NULL, ignore_case)
+  select <- .select_nse(
+    select,
+    data,
+    exclude = NULL,
+    ignore_case = ignore_case,
+    regex = regex,
+    verbose = FALSE)
 
   # nothing to remove?
   if (!length(select)) {
