+++ man/data_relocate.Rd
@@ -18,7 +18,15 @@ data_relocate(
 
 data_reorder(data, select, ignore_case = FALSE, verbose = TRUE, ...)
 
-data_remove(data, select, ignore_case = FALSE, verbose = FALSE, ...)
+data_remove(
+  data,
+  select = NULL,
+  exclude = NULL,
+  ignore_case = FALSE,
+  regex = FALSE,
+  verbose = FALSE,
+  ...
+)
 }
 \arguments{
 \item{data}{A data frame.}
@@ -66,6 +74,19 @@ search pattern when matching against variable names.}
 \item{verbose}{Toggle warnings.}
 
 \item{...}{Arguments passed down to other functions. Mostly not used yet.}
+
+\item{exclude}{See \code{select}, however, column names matched by the pattern
+from \code{exclude} will be excluded instead of selected. If \code{NULL} (the default),
+excludes no columns.}
+
+\item{regex}{Logical, if \code{TRUE}, the search pattern from \code{select} will be
+treated as regular expression. When \code{regex = TRUE}, select \emph{must} be a
+character string (or a variable containing a character string) and is not
+allowed to be one of the supported select-helpers or a character vector
+of length > 1. \code{regex = TRUE} is comparable to using one of the two
+select-helpers, \code{select = contains("")} or \code{select = regex("")}, however,
+since the select-helpers may not work when called from inside other
+functions (see 'Details'), this argument may be used as workaround.}
 }
 \value{
 A data frame with reordered columns.
