### 2022-08-24 2ddb555
added regex argument (+ tests) to all the other frigging functions
