+++ R/data_tabulate.R
@@ -130,12 +130,18 @@ data_tabulate.data.frame <- function(x,
                                      select = NULL,
                                      exclude = NULL,
                                      ignore_case = FALSE,
+                                     regex = FALSE,
                                      collapse = FALSE,
                                      drop_levels = FALSE,
                                      verbose = TRUE,
                                      ...) {
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = FALSE)
   out <- lapply(select, function(i) {
     data_tabulate(x[[i]], drop_levels = drop_levels, name = i, verbose = verbose, ...)
   })
@@ -152,6 +158,7 @@ data_tabulate.grouped_df <- function(x,
                                      select = NULL,
                                      exclude = NULL,
                                      ignore_case = FALSE,
+                                     regex = FALSE,
                                      verbose = TRUE,
                                      collapse = FALSE,
                                      drop_levels = FALSE,
@@ -161,7 +168,12 @@ data_tabulate.grouped_df <- function(x,
   group_variables <- NULL
 
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = FALSE)
 
   # dplyr < 0.8.0?
   if (is.null(grps)) {
