+++ R/slide.R
@@ -62,10 +62,16 @@ slide.data.frame <- function(x,
                              lowest = 0,
                              append = FALSE,
                              ignore_case = FALSE,
+                             regex = FALSE,
                              verbose = TRUE,
                              ...) {
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   # process arguments
   args <- .process_std_args(
