### 2022-08-24 b1462b0
changed verbose to use default instead of forced FALSE deeper within code
