+++ R/data_tabulate.R
@@ -141,7 +141,7 @@ data_tabulate.data.frame <- function(x,
                         exclude,
                         ignore_case,
                         regex = regex,
-                        verbose = FALSE)
+                        verbose = verbose)
   out <- lapply(select, function(i) {
     data_tabulate(x[[i]], drop_levels = drop_levels, name = i, verbose = verbose, ...)
   })
@@ -173,7 +173,7 @@ data_tabulate.grouped_df <- function(x,
                         exclude,
                         ignore_case,
                         regex = regex,
-                        verbose = FALSE)
+                        verbose = verbose)
 
   # dplyr < 0.8.0?
   if (is.null(grps)) {
