+++ R/data_reverse.R
@@ -145,7 +145,7 @@ reverse.grouped_df <- function(x,
                         exclude,
                         ignore_case,
                         regex = regex,
-                        verbose = FALSE)
+                        verbose = verbose)
 
   # dplyr < 0.8.0?
   if (is.null(grps)) {
@@ -188,7 +188,7 @@ reverse.data.frame <- function(x,
                         exclude,
                         ignore_case,
                         regex = regex,
-                        verbose = FALSE)
+                        verbose = verbose)
 
   # Transform the range so that it is a list now
   if (!is.null(range)) {
