### 2022-08-24 5be46db
added tests with "real" regex pattern
