+++ tests/testthat/test-adjust.R
@@ -22,5 +22,8 @@ test_that("adjust regex", {
   expect_equal(
     adjust(mtcars, select = "pg", regex = TRUE),
     adjust(mtcars, select = "mpg"))
+  expect_equal(
+    adjust(mtcars, select = "pg$", regex = TRUE),
+    adjust(mtcars, select = "mpg")
+  )
 })
-
