+++ tests/testthat/test-center.R
@@ -132,4 +132,8 @@ test_that("center regex", {
   expect_equal(
     center(mtcars, select = "pg", regex = TRUE)$mpg,
     center(mtcars$mpg))
+  expect_equal(
+    center(mtcars, select = "pg$", regex = TRUE)$mpg,
+    center(mtcars$mpg)
+  )
 })
