### 2022-08-24 03d6bcb
add regex argument to all functions (#230)
* add regex argument to data_remove

* added regex argument (+ tests) to all the other frigging functions

* changed verbose to use default instead of forced FALSE deeper within code

* added tests with "real" regex pattern

* news

Co-authored-by: Daniel <mail@danielluedecke.de>
