+++ R/convert_na_to.R
@@ -154,10 +154,16 @@ convert_na_to.data.frame <- function(x,
                                      replace_char = replacement,
                                      replace_fac = replacement,
                                      ignore_case = FALSE,
+                                     regex = FALSE,
                                      verbose = TRUE,
                                      ...) {
   data <- x
-  select_nse <- .select_nse(select, data, exclude = exclude, ignore_case)
+  select_nse <- .select_nse(select,
+                            data,
+                            exclude = exclude,
+                            ignore_case,
+                            regex = regex,
+                            verbose = verbose)
 
   # list are not covered by .select_nse
   if (length(select_nse) == 0) {
