+++ R/to_factor.R
@@ -83,6 +83,7 @@ to_factor.data.frame <- function(x,
                                  exclude = NULL,
                                  ignore_case = FALSE,
                                  append = FALSE,
+                                 regex = FALSE,
                                  verbose = TRUE,
                                  ...) {
   # sanity check, return as is for complete numeric
@@ -91,7 +92,12 @@ to_factor.data.frame <- function(x,
   }
 
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   # drop factors, when append is not FALSE
   if (!isFALSE(append)) {
