+++ tests/testthat/test-data_extract.R
@@ -131,3 +131,13 @@ test_that("data_extract extract, pull", {
     c("Sepal.Length", "Sepal.Width", "Species")
   )
 })
+
+# select helpers ------------------------------
+test_that("data_extract regex", {
+  expect_equal(
+    data_extract(mtcars, select = "pg", regex = TRUE),
+    data_extract(mtcars, select = "mpg"))
+  expect_equal(
+    data_extract(mtcars, select = "pg$", regex = TRUE),
+    data_extract(mtcars, select = "mpg"))
+})
