+++ tests/testthat/test-unnormalize.R
@@ -39,6 +39,6 @@ test_that("unnormalize regex", {
   x <- normalize(mtcars, select = "mpg")
   expect_equal(
     unnormalize(x, select = "pg", regex = TRUE),
-    unnormalize(x, select = "mpg"))
+    unnormalize(x, select = "mpg")
+  )
 })
-
