+++ WIP/data_recode_strange.R
@@ -1,5 +1,5 @@
 #' @title Recode old values of variables into new values
-#' @name change_code
+#' @name recode_values
 #'
 #' @description
 #' This functions recodes old values into new values and can be used to to
@@ -81,12 +81,12 @@
 #' x <- sample(c(1:4, NA), 15, TRUE)
 #' table(x, useNA = "always")
 #'
-#' out <- change_code(x, list(`0` = 1, `1` = 2:3, `2` = 4))
+#' out <- recode_values(x, list(`0` = 1, `1` = 2:3, `2` = 4))
 #' out
 #' table(out, useNA = "always")
 #'
 #' # to recode NA values, set preserve_na to FALSE
-#' out <- change_code(
+#' out <- recode_values(
 #'   x,
 #'   list(`0` = 1, `1` = 2:3, `2` = 4, `9` = NA),
 #'   preserve_na = FALSE
@@ -95,12 +95,12 @@
 #' table(out, useNA = "always")
 #'
 #' # preserve na
-#' out <- change_code(x, list(`0` = 1, `1` = 2:3), default = 77)
+#' out <- recode_values(x, list(`0` = 1, `1` = 2:3), default = 77)
 #' out
 #' table(out, useNA = "always")
 #'
 #' # recode na into default
-#' out <- change_code(
+#' out <- recode_values(
 #'   x,
 #'   list(`0` = 1, `1` = 2:3),
 #'   default = 77,
@@ -115,17 +115,17 @@
 #' x <- as.factor(sample(c("a", "b", "c"), 15, TRUE))
 #' table(x)
 #'
-#' out <- change_code(x, list(x = "a", y = c("b", "c")))
+#' out <- recode_values(x, list(x = "a", y = c("b", "c")))
 #' out
 #' table(out)
 #'
-#' out <- change_code(x, list(x = "a", y = "b", z = "c"))
+#' out <- recode_values(x, list(x = "a", y = "b", z = "c"))
 #' out
 #' table(out)
 #'
-#' out <- change_code(x, list(y = "b,c"), default = 77)
+#' out <- recode_values(x, list(y = "b,c"), default = 77)
 #' # same as
-#' # change_code(x, list(y = c("b", "c")), default = 77)
+#' # recode_values(x, list(y = c("b", "c")), default = 77)
 #' out
 #' table(out)
 #'
@@ -138,20 +138,20 @@
 #'   stringsAsFactors = FALSE
 #' )
 #'
-#' change_code(
+#' recode_values(
 #'   d,
 #'   recodes = list(`0` = 1, `1` = 2:3, `2` = 4, x = "a", y = c("b", "c")),
 #'   force = TRUE,
 #'   append = TRUE
 #' )
 #' @export
-change_code <- function(x, ...) {
-  UseMethod("change_code")
+recode_values <- function(x, ...) {
+  UseMethod("recode_values")
 }
 
 
 #' @export
-change_code.default <- function(x, verbose = TRUE, ...) {
+recode_values.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
     message(insight::format_message(paste0("Variables of class '", class(x)[1], "' can't be recoded and remain unchanged.")))
   }
@@ -159,14 +159,14 @@ change_code.default <- function(x, verbose = TRUE, ...) {
 }
 
 
-#' @rdname change_code
+#' @rdname recode_values
 #' @export
-change_code.numeric <- function(x,
-                                recodes = NULL,
-                                default = NULL,
-                                preserve_na = TRUE,
-                                verbose = TRUE,
-                                ...) {
+recode_values.numeric <- function(x,
+                                  recodes = NULL,
+                                  default = NULL,
+                                  preserve_na = TRUE,
[…]
