+++ tests/testthat/test-ranktransform.R
@@ -113,6 +113,6 @@ test_that("ranktransform works with data frames containing NAs (grouped data)", 
 test_that("ranktransform regex", {
   expect_equal(
     ranktransform(mtcars, select = "pg", regex = TRUE),
-    ranktransform(mtcars, select = "mpg"))
+    ranktransform(mtcars, select = "mpg")
+  )
 })
-
