+++ R/unnormalize.R
@@ -46,11 +46,12 @@ unnormalize.data.frame <- function(x,
                                    ...) {
   # evaluate select/exclude, may be select-helpers
   select <- .select_nse(select,
-                        x,
-                        exclude,
-                        ignore_case,
-                        regex = regex,
-                        verbose = verbose)
+    x,
+    exclude,
+    ignore_case,
+    regex = regex,
+    verbose = verbose
+  )
   x[select] <- lapply(x[select], unnormalize, verbose = verbose)
 
   x

[…]
