### 2022-08-24 6adeff4
add lifecycle badge
https://github.com/easystats/easystats/issues/271
