+++ R/recode_values.R
@@ -1,5 +1,5 @@
 #' @title Recode old values of variables into new values
-#' @name change_code
+#' @name recode_values
 #'
 #' @description
 #' This functions recodes old values into new values and can be used to to
@@ -89,12 +89,12 @@
 #' x <- sample(c(1:4, NA), 15, TRUE)
 #' table(x, useNA = "always")
 #'
-#' out <- change_code(x, list(`0` = 1, `1` = 2:3, `2` = 4))
+#' out <- recode_values(x, list(`0` = 1, `1` = 2:3, `2` = 4))
 #' out
 #' table(out, useNA = "always")
 #'
 #' # to recode NA values, set preserve_na to FALSE
-#' out <- change_code(
+#' out <- recode_values(
 #'   x,
 #'   list(`0` = 1, `1` = 2:3, `2` = 4, `9` = NA),
 #'   preserve_na = FALSE
@@ -103,12 +103,12 @@
 #' table(out, useNA = "always")
 #'
 #' # preserve na ----------
-#' out <- change_code(x, list(`0` = 1, `1` = 2:3), default = 77)
+#' out <- recode_values(x, list(`0` = 1, `1` = 2:3), default = 77)
 #' out
 #' table(out, useNA = "always")
 #'
 #' # recode na into default ----------
-#' out <- change_code(
+#' out <- recode_values(
 #'   x,
 #'   list(`0` = 1, `1` = 2:3),
 #'   default = 77,
@@ -123,17 +123,17 @@
 #' x <- as.factor(sample(c("a", "b", "c"), 15, TRUE))
 #' table(x)
 #'
-#' out <- change_code(x, list(x = "a", y = c("b", "c")))
+#' out <- recode_values(x, list(x = "a", y = c("b", "c")))
 #' out
 #' table(out)
 #'
-#' out <- change_code(x, list(x = "a", y = "b", z = "c"))
+#' out <- recode_values(x, list(x = "a", y = "b", z = "c"))
 #' out
 #' table(out)
 #'
-#' out <- change_code(x, list(y = "b,c"), default = 77)
+#' out <- recode_values(x, list(y = "b,c"), default = 77)
 #' # same as
-#' # change_code(x, list(y = c("b", "c")), default = 77)
+#' # recode_values(x, list(y = c("b", "c")), default = 77)
 #' out
 #' table(out)
 #'
@@ -146,7 +146,7 @@
 #'   stringsAsFactors = FALSE
 #' )
 #'
-#' change_code(
+#' recode_values(
 #'   d,
 #'   recode = list(`0` = 1, `1` = 2:3, `2` = 4, x = "a", y = c("b", "c")),
 #'   append = TRUE
@@ -161,7 +161,7 @@
 #' x <- sample(c(1:4, NA), 15, TRUE)
 #' table(x, useNA = "always")
 #'
-#' out <- change_code(x, list(`1` = 0, `2:3` = 1, `4` = 2))
+#' out <- recode_values(x, list(`1` = 0, `2:3` = 1, `4` = 2))
 #' table(out, useNA = "always")
 #'
 #' # factors (character vectors are similar)
@@ -169,19 +169,19 @@
 #' x <- as.factor(sample(c("a", "b", "c"), 15, TRUE))
 #' table(x)
 #'
-#' out <- change_code(x, list(a = "x", `b, c` = "y"))
+#' out <- recode_values(x, list(a = "x", `b, c` = "y"))
 #' table(out)
 #'
 #' # reset options
 #' options(data_recode_pattern = NULL)
 #' @export
-change_code <- function(x, ...) {
-  UseMethod("change_code")
+recode_values <- function(x, ...) {
+  UseMethod("recode_values")
 }
 
 
 #' @export
[…]
