+++ tests/testthat/test-data_extract.R
@@ -136,8 +136,10 @@ test_that("data_extract extract, pull", {
 test_that("data_extract regex", {
   expect_equal(
     data_extract(mtcars, select = "pg", regex = TRUE),
-    data_extract(mtcars, select = "mpg"))
+    data_extract(mtcars, select = "mpg")
+  )
   expect_equal(
     data_extract(mtcars, select = "pg$", regex = TRUE),
-    data_extract(mtcars, select = "mpg"))
+    data_extract(mtcars, select = "mpg")
+  )
 })
