+++ tests/testthat/test-convert_na_to.R
@@ -413,7 +413,8 @@ test_that("data_rename preserves attributes", {
 test_that("convert_na_to regex", {
   expect_equal(
     convert_na_to(airquality, replacement = 0, select = "zone", regex = TRUE),
-    convert_na_to(airquality, replacement = 0, select = "Ozone"))
+    convert_na_to(airquality, replacement = 0, select = "Ozone")
+  )
   expect_equal(
     convert_na_to(airquality, replacement = 0, select = "zone$", regex = TRUE),
     convert_na_to(airquality, replacement = 0, select = "Ozone")
