+++ tests/testthat/test-convert_to_na.R
@@ -95,7 +95,8 @@ test_that("convert_to_na other classes", {
 test_that("convert_to_na regex", {
   expect_equal(
     convert_to_na(mtcars, na = 4, select = "arb", regex = TRUE),
-    convert_to_na(mtcars, na = 4, select = "carb"))
+    convert_to_na(mtcars, na = 4, select = "carb")
+  )
   expect_equal(
     convert_to_na(mtcars, na = 4, select = "arb$", regex = TRUE),
     convert_to_na(mtcars, na = 4, select = "carb")
