+++ _pkgdown.yml
@@ -100,4 +100,3 @@ articles:
     navbar: ~
     contents:
       - standardize_data
-
