+++ .github/workflows/pkgdown.yaml
@@ -21,7 +21,7 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
         with:
-            pandoc-version: '2.19.1'
+            pandoc-version: '2.19.2'
 
       - uses: r-lib/actions/setup-r@v2
         with:
@@ -29,8 +29,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: |
-              local::.
+          extra-packages: r-lib/pkgdown, local::.
           needs: website
 
       - name: Build site
