### 2022-08-28 3ca9b46
Add coefficient of variation function
Related to https://github.com/easystats/datawizard/issues/232, https://github.com/easystats/performance/issues/433
