+++ man/coef_var.Rd
@@ -1,0 +1,72 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/describe_distribution.R
+\name{coef_var}
+\alias{coef_var}
+\alias{cv}
+\alias{distribution_cv}
+\alias{coef_var.numeric}
+\title{Compute the coefficient of variation}
+\usage{
+coef_var(x, ...)
+
+cv(x, ...)
+
+distribution_cv(x, ...)
+
+\method{coef_var}{numeric}(
+  x,
+  mu = NULL,
+  sigma = NULL,
+  unbiased = TRUE,
+  method = c("standard", "unbiased", "median_mad", "qcd"),
+  trim = 0,
+  na.rm = FALSE,
+  n = NULL,
+  ...
+)
+}
+\arguments{
+\item{x}{A numeric vector, or a model of a supported class.}
+
+\item{...}{Further arguments based to other methods.}
+
+\item{mu}{A numeric vector of mean values to use to compute the coefficient
+of variation. If supplied, \code{x} is not used to compute the mean.}
+
+\item{sigma}{A numeric vector of standard deviation values to use to compute the coefficient
+of variation. If supplied, \code{x} is not used to compute the SD.}
+
+\item{method}{Method to use to compute the CV. Can be \code{"standard"} to compute
+by dividing the standard deviation by the mean, \code{"unbiased"} for the
+unbiased estimator for normally distributed data, or one of two robust
+alternatives: \code{"median_mad"} to divide the median by the \code{\link[stats:mad]{stats::mad()}},
+or \code{"qcd"} (quartile coefficient of dispersion, interquartile range divided
+by the sum of the quartiles \link{twice the midhinge}: $\frac{Q_3 - Q_1}{Q_3 + Q_1}$).}
+
+\item{trim}{the fraction (0 to 0.5) of values to be trimmed from
+each end of \code{x} before the mean and standard deviation (or alternatvies)
+are computed. Values of \code{trim} outside the range of (0 to 0.5) are taken
+as the nearest endpoint.}
+
+\item{na.rm}{Logical. Should \code{NA} values be removed before computing (\code{TRUE})
+or not (\code{FALSE}, default)?}
+
+\item{n}{If \code{method = "unbiased"} and both \code{mu} and \code{sigma} are provided (not
+computed from \code{x}), what sample size to use to adjust the computed CV
+for small-sample bias?}
+}
+\value{
+The computed coefficient of variation for \code{x}.
+}
+\description{
+Compute the coefficient of variation (CV, ratio of the standard deviation to
+the mean, $\frac{\sigma}{\mu}$) for a set of numeric values. Can also be
+used to compute the CV for a fitted model.
+}
+\examples{
+coef_var(1:10)
+coef_var(1:10, method = "qcd")
+coef_var(mu = 10, sigma = 20)
+coef_var(mu = 10, sigma = 20, method = "unbiased", n = 30)
+cv(1:10)
+}
