+++ NAMESPACE
@@ -20,6 +20,8 @@ S3method(center,factor)
 S3method(center,grouped_df)
 S3method(center,logical)
 S3method(center,numeric)
+S3method(coef_var,default)
+S3method(coef_var,numeric)
 S3method(convert_na_to,character)
 S3method(convert_na_to,data.frame)
 S3method(convert_na_to,default)
@@ -158,6 +160,7 @@ export(center)
 export(centre)
 export(change_code)
 export(change_scale)
+export(coef_var)
 export(coerce_to_numeric)
 export(colnames_to_row)
 export(column_as_rownames)
@@ -165,6 +168,7 @@ export(compact_character)
 export(compact_list)
 export(convert_na_to)
 export(convert_to_na)
+export(cv)
 export(data_addprefix)
 export(data_addsuffix)
 export(data_adjust)
@@ -195,6 +199,7 @@ export(degroup)
 export(demean)
 export(describe_distribution)
 export(detrend)
+export(distribution_cv)
 export(distribution_mode)
 export(empty_columns)
 export(empty_rows)
