### 2022-08-28 9be7427
Remove suggestion that `coef_var()` can be used with models.
Will implement that as `performance_coefvar()` or similar
