+++ R/describe_distribution.R
@@ -538,10 +538,9 @@ distribution_mode <- function(x) {
 #' Compute the coefficient of variation
 #'
 #' Compute the coefficient of variation (CV, ratio of the standard deviation to
-#' the mean, $\frac{\sigma}{\mu}$) for a set of numeric values. Can also be
-#' used to compute the CV for a fitted model.
+#' the mean, $\frac{\sigma}{\mu}$) for a set of numeric values.
 #'
-#' @param x A numeric vector, or a model of a supported class.
+#' @param x A numeric vector, or vector of values than can be converted to numeric.
 #' @param ... Further arguments based to other methods.
 #'
 #' @return The computed coefficient of variation for `x`.
