### 2022-08-29 c5cfb89
need to catch exceptions
