+++ tests/testthat/test-data_match.R
@@ -78,3 +78,22 @@ test_that("data_filter works like slice", {
   out <- data_filter(mtcars, 5:10)
   expect_equal(out, mtcars[5:10, ], ignore_attr = TRUE)
 })
+
+test_that("data_filter gives informative message on errors", {
+  expect_error(
+    data_filter(mtcars, "filter = mpg > 10 || cyl = 4"),
+    "`==`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 || cyl == 4),
+    "`||`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 && cyl == 4),
+    "`&&`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 ? cyl == 4),
+    "syntax"
+  )
+})
