### 2022-08-29 cdc9818
Use Markdown in .Rd
