### 2022-08-29 4f94216
Move mode and CV to descriptives.R
