+++ R/describe_distribution.R
@@ -505,146 +505,3 @@ print.parameters_distribution <- function(x, digits = 2, ...) {
   )
   out[[1]]
 }
-
-# distribution_mode ----------------------------------
-
-#' Compute mode for a statistical distribution
-#'
-#' @param x An atomic vector, a list, or a data frame.
-#'
-#' @return
-#'
-#' The value that appears most frequently in the provided data.
-#' The returned data structure will be the same as the entered one.
-#'
-#' @seealso For continuous variables, the
-#'   **Highest Maximum a Posteriori probability estimate (MAP)** may be
-#'   a more useful way to estimate the most commonly-observed value
-#'   than the mode. See [bayestestR::map_estimate()].
-#'
-#' @examples
-#'
-#' distribution_mode(c(1, 2, 3, 3, 4, 5))
-#' distribution_mode(c(1.5, 2.3, 3.7, 3.7, 4.0, 5))
-#'
-#' @export
-distribution_mode <- function(x) {
-  uniqv <- unique(x)
-  tab <- tabulate(match(x, uniqv))
-  idx <- which.max(tab)
-  uniqv[idx]
-}
-
-#' Compute the coefficient of variation
-#'
-#' Compute the coefficient of variation (CV, ratio of the standard deviation to
-#' the mean, $\frac{\sigma}{\mu}$) for a set of numeric values.
-#'
-#' @param x A numeric vector, or vector of values than can be converted to numeric.
-#' @param ... Further arguments based to other methods.
-#'
-#' @return The computed coefficient of variation for `x`.
-#' @export
-#'
-#' @examples
-#' coef_var(1:10)
-#' coef_var(1:10, method = "qcd")
-#' coef_var(mu = 10, sigma = 20)
-#' coef_var(mu = 10, sigma = 20, method = "unbiased", n = 30)
-#' cv(1:10)
-coef_var <- function(x, ...) {
-  UseMethod("coef_var")
-}
-
-#' @name cv
-#' @rdname coef_var
-#' @export
-cv <- coef_var
-
-#' @name distribution_cv
-#' @rdname coef_var
-#' @export
-distribution_cv <- coef_var
-
-#' @export
-coef_var.default <- function(x, verbose = TRUE, ...) {
-  if (verbose) {
-    warning(insight::format_message(
-      paste0("Can't compute the coefficient of variation objects of class '", class(x)[1], "'.")
-    ), call. = FALSE)
-  }
-  NULL
-}
-
-#' @param mu A numeric vector of mean values to use to compute the coefficient
-#'   of variation. If supplied, `x` is not used to compute the mean.
-#' @param sigma A numeric vector of standard deviation values to use to compute the coefficient
-#'   of variation. If supplied, `x` is not used to compute the SD.
-#' @param method Method to use to compute the CV. Can be `"standard"` to compute
-#'   by dividing the standard deviation by the mean, `"unbiased"` for the
-#'   unbiased estimator for normally distributed data, or one of two robust
-#'   alternatives: `"median_mad"` to divide the median by the [stats::mad()],
-#'   or `"qcd"` (quartile coefficient of dispersion, interquartile range divided
-#'   by the sum of the quartiles [twice the midhinge]: $\frac{Q_3 - Q_1}{Q_3 + Q_1}$).
-#' @param trim the fraction (0 to 0.5) of values to be trimmed from
-#'   each end of `x` before the mean and standard deviation (or alternatvies)
-#'   are computed. Values of `trim` outside the range of (0 to 0.5) are taken
-#'   as the nearest endpoint.
[…]
