### 2022-08-29 ac39f96
Bugfix in switch()
