### 2022-08-29 98d4752
bug fixes and add tests
