### 2022-08-29 cc84f46
Move processing to internal functions for reuse
