+++ man/coef_var.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/describe_distribution.R
+% Please edit documentation in R/descriptives.R
 \name{coef_var}
 \alias{coef_var}
 \alias{cv}
@@ -27,7 +27,7 @@ distribution_cv(x, ...)
 \arguments{
 \item{x}{A numeric vector, or vector of values than can be converted to numeric.}
 
-\item{...}{Further arguments based to other methods.}
+\item{...}{Further arguments passed to computation functions.}
 
 \item{mu}{A numeric vector of mean values to use to compute the coefficient
 of variation. If supplied, \code{x} is not used to compute the mean.}
@@ -40,7 +40,7 @@ by dividing the standard deviation by the mean, \code{"unbiased"} for the
 unbiased estimator for normally distributed data, or one of two robust
 alternatives: \code{"median_mad"} to divide the median by the \code{\link[stats:mad]{stats::mad()}},
 or \code{"qcd"} (quartile coefficient of dispersion, interquartile range divided
-by the sum of the quartiles \link{twice the midhinge}: $\frac{Q_3 - Q_1}{Q_3 + Q_1}$).}
+by the sum of the quartiles [twice the midhinge]: \eqn{(Q_3 - Q_1)/(Q_3 + Q_1)}.}
 
 \item{trim}{the fraction (0 to 0.5) of values to be trimmed from
 each end of \code{x} before the mean and standard deviation (or alternatvies)
@@ -59,11 +59,12 @@ The computed coefficient of variation for \code{x}.
 }
 \description{
 Compute the coefficient of variation (CV, ratio of the standard deviation to
-the mean, $\frac{\sigma}{\mu}$) for a set of numeric values.
+the mean, \eqn{\sigma/\mu}) for a set of numeric values.
 }
 \examples{
 coef_var(1:10)
-coef_var(1:10, method = "qcd")
+coef_var(c(1:10, 100), method = "median_mad")
+coef_var(c(1:10, 100), method = "qcd")
 coef_var(mu = 10, sigma = 20)
 coef_var(mu = 10, sigma = 20, method = "unbiased", n = 30)
 cv(1:10)
