+++ man/distribution_mode.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/describe_distribution.R
+% Please edit documentation in R/descriptives.R
 \name{distribution_mode}
 \alias{distribution_mode}
 \title{Compute mode for a statistical distribution}
@@ -25,6 +25,6 @@ distribution_mode(c(1.5, 2.3, 3.7, 3.7, 4.0, 5))
 \seealso{
 For continuous variables, the
 \strong{Highest Maximum a Posteriori probability estimate (MAP)} may be
-more a more useful way to estimate the most commonly-observed value
+a more useful way to estimate the most commonly-observed value
 than the mode. See \code{\link[bayestestR:map_estimate]{bayestestR::map_estimate()}}.
 }
