+++ .github/workflows/R-CMD-check.yaml
@@ -19,21 +19,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          #- {os: macOS-latest,   r: 'devel'}
-          - {os: macOS-latest,   r: 'release'}
           - {os: macOS-latest,   r: 'oldrel'}
-
-          - {os: windows-latest, r: 'devel'}
-          - {os: windows-latest, r: 'release'}
-          - {os: windows-latest, r: 'oldrel'}
-          - {os: windows-latest, r: 'oldrel-2'}
-          #- {os: windows-latest, r: '3.6.0'}
-          #- {os: windows-latest, r: '3.5.0'}
-
-          - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
-          - {os: ubuntu-latest,  r: 'release'}
-          - {os: ubuntu-latest,  r: 'oldrel'}
-          #- {os: ubuntu-latest,   r: '3.6.0'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
