### 2022-08-30 8478a94
try to get more info for debugging
