### 2022-08-30 20c12de
use try instead of tryCatch (#237)
