+++ tests/testthat/test-data_match.R
@@ -80,7 +80,6 @@ test_that("data_filter works like slice", {
 })
 
 test_that("data_filter gives informative message on errors", {
-  data(mtcars)
   expect_error(
     data_filter(mtcars, "mpg > 10 || cyl = 4"),
     "`==`"
