### 2022-08-30 e4f8e01
Deploying to gh-pages from @ easystats/datawizard@20c12de74d0a9fdb39b5f92ca30923e34a68c052 🚀
