### 2022-08-30 68d1a1e
use tryCatch again, to capture warnings as well
