### 2022-08-30 55d710a
Deploying to gh-pages from @ easystats/datawizard@68d1a1e6c73dd968f1f4f08f5c651cd8b55351f1 🚀
