### 2022-08-30 59b8006
Deploying to gh-pages from @ easystats/datawizard@fe8d44b3600189dd857df58a5d4ab88b2b719d0e 🚀
