### 2022-08-30 7cd9307
move filter check before calling subset()
