+++ tests/testthat/test-data_match.R
@@ -81,21 +81,21 @@ test_that("data_filter works like slice", {
 
 ## TODO These tests doesn't work from test-environment
 
-# test_that("data_filter gives informative message on errors", {
-#   expect_error(
-#     data_filter(mtcars, "mpg > 10 || cyl = 4"),
-#     "`==`"
-#   )
-#   expect_error(
-#     data_filter(mtcars, filter = mpg > 10 || cyl == 4),
-#     "`||`"
-#   )
-#   expect_error(
-#     data_filter(mtcars, filter = mpg > 10 && cyl == 4),
-#     "`&&`"
-#   )
-#   expect_error(
-#     data_filter(mtcars, filter = mpg > 10 ? cyl == 4),
-#     "syntax"
-#   )
-# })
+test_that("data_filter gives informative message on errors", {
+  expect_error(
+    data_filter(mtcars, "mpg > 10 || cyl = 4"),
+    "`==`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 || cyl == 4),
+    "`||`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 && cyl == 4),
+    "`&&`"
+  )
+  expect_error(
+    data_filter(mtcars, filter = mpg > 10 ? cyl == 4),
+    "syntax"
+  )
+})
