### 2022-09-01 d89c25e
Use new GHA workflows for easystats
https://github.com/easystats/easystats/issues/234
