+++ man/data_rename.Rd
@@ -13,6 +13,8 @@ data_addprefix(
   select = NULL,
   exclude = NULL,
   ignore_case = FALSE,
+  regex = FALSE,
+  verbose = TRUE,
   ...
 )
 
@@ -74,6 +76,17 @@ excludes no columns.}
 \item{ignore_case}{Logical, if \code{TRUE} and when one of the select-helpers or
 a regular expression is used in \code{select}, ignores lower/upper case in the
 search pattern when matching against variable names.}
+
+\item{regex}{Logical, if \code{TRUE}, the search pattern from \code{select} will be
+treated as regular expression. When \code{regex = TRUE}, select \emph{must} be a
+character string (or a variable containing a character string) and is not
+allowed to be one of the supported select-helpers or a character vector
+of length > 1. \code{regex = TRUE} is comparable to using one of the two
+select-helpers, \code{select = contains("")} or \code{select = regex("")}, however,
+since the select-helpers may not work when called from inside other
+functions (see 'Details'), this argument may be used as workaround.}
+
+\item{verbose}{Toggle warnings.}
 
 \item{...}{Other arguments passed to or from other functions.}
 
