### 2022-09-01 fd1dc14
allow <= and >= in data_filter
