### 2022-09-01 1b96052
add regex to `data_addprefix` for correct use in `performance::check_outliers` (#240)
* add regex to data_addprefix for correct use in performance::check_outliers

* corrected test/oversight

Co-authored-by: Daniel <mail@danielluedecke.de>
