+++ R/data_addprefix.R
@@ -6,9 +6,21 @@
 #' head(data_addsuffix(iris, "_OLD"))
 #'
 #' @export
-data_addprefix <- function(data, pattern, select = NULL, exclude = NULL, ignore_case = FALSE, ...) {
+data_addprefix <- function(data,
+                           pattern,
+                           select = NULL,
+                           exclude = NULL,
+                           ignore_case = FALSE,
+                           regex = FALSE,
+                           verbose = TRUE,
+                           ...) {
   # evaluate arguments
-  select <- .select_nse(select, data, exclude, ignore_case)
+  select <- .select_nse(select,
+                        data,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   selected_columns <- colnames(data) %in% select
   colnames(data)[selected_columns] <- paste0(pattern, colnames(data)[selected_columns])
@@ -18,9 +30,21 @@ data_addprefix <- function(data, pattern, select = NULL, exclude = NULL, ignore_
 
 #' @rdname data_rename
 #' @export
-data_addsuffix <- function(data, pattern, select = NULL, exclude = NULL, ignore_case = FALSE, ...) {
+data_addsuffix <- function(data,
+                           pattern,
+                           select = NULL,
+                           exclude = NULL,
+                           ignore_case = FALSE,
+                           regex = FALSE,
+                           verbose = TRUE,
+                           ...) {
   # evaluate arguments
-  select <- .select_nse(select, data, exclude, ignore_case)
+  select <- .select_nse(select,
+                        data,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   selected_columns <- colnames(data) %in% select
   colnames(data)[selected_columns] <- paste0(colnames(data)[selected_columns], pattern)
