### 2022-09-01 362bf43
data_filter: allow <= and >=  (#242)
* allow <= and >= in data_filter

* typo

* more tests

* typo again...

Co-authored-by: Daniel <mail@danielluedecke.de>
