+++ tests/testthat/test-data_match.R
@@ -97,3 +97,31 @@ test_that("data_filter gives informative message on errors", {
     "syntax"
   )
 })
+
+test_that("data_filter works with >= or <=", {
+  expect_equal(
+    data_filter(mtcars, "mpg >= 30.4"),
+    subset(mtcars, mpg >= 30.4)
+  )
+  expect_equal(
+    data_filter(mtcars, mpg >= 30.4),
+    subset(mtcars, mpg >= 30.4)
+  )
+  expect_equal(
+    data_filter(mtcars, "mpg <= 30.4"),
+    subset(mtcars, mpg <= 30.4)
+  )
+  expect_equal(
+    data_filter(mtcars, mpg <= 30.4),
+    subset(mtcars, mpg <= 30.4)
+  )
+
+  expect_equal(
+    data_filter(mtcars, "mpg >= 30.4 & hp == 66"),
+    subset(mtcars, mpg >= 30.4 & hp == 66)
+  )
+  expect_equal(
+    data_filter(mtcars, mpg <= 30.4 & hp == 66),
+    subset(mtcars, mpg <= 30.4 & hp == 66)
+  )
+})
