### 2022-09-01 0e59551
Deploying to gh-pages from @ easystats/datawizard@362bf437202164f89f9d0fbe51f24ffaf85fd90e 🚀
