### 2022-09-02 cc3c861
typos in news
