+++ NAMESPACE
@@ -35,6 +35,7 @@ S3method(convert_to_na,numeric)
 S3method(data_extract,data.frame)
 S3method(data_merge,data.frame)
 S3method(data_merge,list)
+S3method(data_peek,data.frame)
 S3method(data_tabulate,data.frame)
 S3method(data_tabulate,default)
 S3method(data_tabulate,grouped_df)
@@ -45,6 +46,7 @@ S3method(describe_distribution,factor)
 S3method(describe_distribution,grouped_df)
 S3method(describe_distribution,list)
 S3method(describe_distribution,numeric)
+S3method(format,dw_data_peek)
 S3method(format,dw_data_tabulate)
 S3method(format,parameters_distribution)
 S3method(kurtosis,data.frame)
@@ -57,6 +59,7 @@ S3method(normalize,grouped_df)
 S3method(normalize,matrix)
 S3method(normalize,numeric)
 S3method(plot,visualisation_recipe)
+S3method(print,dw_data_peek)
 S3method(print,dw_data_tabulate)
 S3method(print,dw_data_tabulates)
 S3method(print,parameters_distribution)
@@ -177,6 +180,7 @@ export(data_join)
 export(data_match)
 export(data_merge)
 export(data_partition)
+export(data_peek)
 export(data_read)
 export(data_relocate)
 export(data_remove)
