+++ NAMESPACE
@@ -66,8 +66,10 @@ S3method(print,parameters_distribution)
 S3method(print,parameters_kurtosis)
 S3method(print,parameters_skewness)
 S3method(print,visualisation_recipe)
+S3method(print_html,dw_data_peek)
 S3method(print_html,dw_data_tabulate)
 S3method(print_html,dw_data_tabulates)
+S3method(print_md,dw_data_peek)
 S3method(print_md,dw_data_tabulate)
 S3method(print_md,dw_data_tabulates)
 S3method(ranktransform,data.frame)
