+++ _pkgdown.yml
@@ -47,6 +47,7 @@ reference:
         Functions to compute statistical summaries of data properties and distributions
     contents:
       - data_tabulate
+      - data_peek
       - contains("distribution")
       - kurtosis
       - smoothness
