### 2022-09-02 ee0d7c3
draft data_peek (#245)
* draft data_peek

* draft finished

* add options

* add options to display md and html

* news

* add test

* pkgdown

* complete docs

* add table caption

* factor is shorter

* create snapshots

Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
