+++ NEWS.md
@@ -13,6 +13,10 @@ BREAKING CHANGES
 * New `recode_values()` alias is introduced for `change_code()`, latter of which
   will be removed in the next release.
 
+NEW FUNCTIONS
+
+* `data_peek()`, to peek at values and type of variables in a data frame.
+
 CHANGES
 
 * `data_filter()` will give more informative messages on malformed syntax of
