+++ tests/testthat/test-data_match.R
@@ -125,3 +125,54 @@ test_that("data_filter works with >= or <=", {
     subset(mtcars, mpg <= 30.4 & hp == 66)
   )
 })
+
+test_that("programming with data_filter", {
+  # One arg ------------
+
+  foo <- function(var) {
+    data_filter(mtcars, filter = {var} > 30)
+  }
+  expect_equal(
+    foo("mpg"),
+    data_filter(mtcars, "mpg >= 30")
+  )
+
+  foo <- function(var) {
+    data_filter(mtcars, filter = "{var} > 30")
+  }
+  expect_equal(
+    foo("mpg"),
+    data_filter(mtcars, "mpg >= 30")
+  )
+
+  # Two args -----------
+
+  foo <- function(var, var2) {
+    data_filter(mtcars, filter = {var} > 30 & {var2} <= 66)
+  }
+  expect_equal(
+    foo("mpg", "hp"),
+    data_filter(mtcars, "mpg >= 30 & hp <= 66")
+  )
+
+  foo <- function(var, var2) {
+    data_filter(mtcars, filter = "{var} > 30 & {var2} <= 66")
+  }
+  expect_equal(
+    foo("mpg", "hp"),
+    data_filter(mtcars, "mpg >= 30 & hp <= 66")
+  )
+})
+
+test_that("programming with data_filter in global env", {
+  var = "mpg"
+  var2 = "hp"
+  expect_equal(
+    data_filter(mtcars, "{var} > 30 & {var2} <= 66"),
+    data_filter(mtcars, "mpg >= 30 & hp <= 66")
+  )
+  expect_equal(
+    data_filter(mtcars, {var} > 30 & {var2} <= 66),
+    data_filter(mtcars, "mpg >= 30 & hp <= 66")
+  )
+})
