+++ .github/workflows/R-CMD-check.yaml
@@ -10,7 +10,7 @@ on:
   pull_request:
     branches: [main, master]
 
-# name: R-CMD-check
+name: R-CMD-check
 
 jobs:
   R-CMD-check:
