### 2022-09-03 224448c
fix tests with R 3.6 (#249)
