### 2022-09-03 fc86356
add news and example in docs
