### 2022-09-03 50b8d26
Enable the use of curly brackets in data_filter() (#217)
* test curley braces

* test curley braces

* add arg eval with curley brace

* add news and example in docs

Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
Co-authored-by: Etienne Bacher <52219252+etiennebacher@users.noreply.github.com>
Co-authored-by: etiennebacher <etienne.bacher@protonmail.com>
