+++ R/data_match.R
@@ -81,7 +81,9 @@
 #' # Define a custom function containing data_filter() and pass variable names
 #' # to it using curly brackets
 #' my_filter <- function(data, variable) {
-#'   data_filter(data, {variable} <= 20)
+#'   data_filter(data, {
+#'     variable
+#'   } <= 20)
 #' }
 #' my_filter(mtcars, "mpg")
 #' @inherit data_rename seealso
@@ -95,12 +97,12 @@ data_match <- function(x, to, match = "and", return_indices = FALSE, drop_na = T
   # evaluate
   match <- match.arg(tolower(match), c("and", "&", "&&", "or", "|", "||", "!", "not"))
   match <- switch(match,
-                  "&" = ,
-                  "&&" = ,
-                  "and" = "and",
-                  "!" = ,
-                  "not" = "not",
-                  "or"
+    "&" = ,
+    "&&" = ,
+    "and" = "and",
+    "!" = ,
+    "not" = "not",
+    "or"
   )
 
   # sanity check
