+++ .github/workflows/lint.yaml
@@ -6,9 +6,8 @@ on:
   pull_request:
     branches: [main, master]
 
-#name: lint
+name: lint
 
 jobs:
   lint:
-    name: lint
     uses: easystats/workflows/.github/workflows/lint.yaml@main
