### 2022-09-04 4242a77
use fixed = TRUE (minimal faster)
