### 2022-09-04 3b37bd9
Deploying to gh-pages from @ easystats/datawizard@931ea56bb49bc50a615ca2ea24b3701c4c92c58c 🚀
