### 2022-09-04 e78ba68
Reduce line length related linters
#247
