### 2022-09-04 86b88e7
Reduce line length related linters (#250)
