+++ tests/testthat/test-std_center.R
@@ -12,7 +12,18 @@ test_that("standardize.data.frame", {
   expect_equal(as.vector(x$Sepal.Length_z), as.vector(scale(d$Sepal.Length)), tolerance = 0.001)
   expect_equal(as.vector(x$Sepal.Length), as.vector(d$Sepal.Length), tolerance = 0.001)
   expect_equal(as.vector(x$Petal.Length), as.vector(d$Petal.Length), tolerance = 0.001)
-  expect_equal(colnames(x), c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width", "Species", "Sepal.Length_z", "Sepal.Width_z"))
+  expect_equal(
+    colnames(x),
+    c(
+      "Sepal.Length",
+      "Sepal.Width",
+      "Petal.Length",
+      "Petal.Width",
+      "Species",
+      "Sepal.Length_z",
+      "Sepal.Width_z"
+    )
+  )
 })
 
 test_that("standardize other classes", {
@@ -25,19 +36,42 @@ test_that("standardize other classes", {
   )
 
   x <- standardize(d$a)
-  expect_equal(as.numeric(x), c(-1.26491, -0.63246, 0, 0.63246, 1.26491), tolerance = 1e-3, ignore_attr = TRUE)
+  expect_equal(
+    as.numeric(x),
+    c(-1.26491, -0.63246, 0, 0.63246, 1.26491),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
   x <- standardize(d$b)
   expect_equal(as.numeric(x), 1:5, tolerance = 1e-3, ignore_attr = TRUE)
   x <- standardize(d$b, force = TRUE)
-  expect_equal(as.numeric(x), c(-1.26491, -0.63246, 0, 0.63246, 1.26491), tolerance = 1e-3, ignore_attr = TRUE)
+  expect_equal(
+    as.numeric(x),
+    c(-1.26491, -0.63246, 0, 0.63246, 1.26491),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
   x <- standardize(d$c)
-  expect_equal(x, as.Date(c("2022-03-22", "2022-01-02", "2022-02-02", "2021-04-02", "2020-01-19")), tolerance = 1e-3, ignore_attr = TRUE)
+  expect_equal(x, as.Date(
+    c(
+      "2022-03-22",
+      "2022-01-02",
+      "2022-02-02",
+      "2021-04-02",
+      "2020-01-19"
+    )
+  ), tolerance = 1e-3, ignore_attr = TRUE)
   x <- standardize(d$c, force = TRUE)
   expect_equal(as.numeric(x), c(0.76992, 0.53121, 0.62488, -0.29975, -1.62626), tolerance = 1e-3, ignore_attr = TRUE)
   x <- standardize(d$d)
   expect_equal(x, c(TRUE, TRUE, FALSE, FALSE, TRUE), tolerance = 1e-3, ignore_attr = TRUE)
   expect_message(x <- standardize(d$d, force = TRUE))
-  expect_equal(x, c(0.7303, 0.7303, -1.09545, -1.09545, 0.7303), tolerance = 1e-3, ignore_attr = TRUE)
+  expect_equal(
+    x,
+    c(0.7303, 0.7303, -1.09545, -1.09545, 0.7303),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
 
   ## TODO conflict with standardize.default() in effectsize
 
@@ -54,10 +88,24 @@ test_that("center.data.frame", {
   expect_equal(colnames(x), c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width", "Species"))
 
[…]
