### 2022-09-04 c1ffaa1
Deploying to gh-pages from @ easystats/datawizard@86b88e75855e6fc0fc09cb916637c9584f9b178d 🚀
