### 2022-09-04 d2828f5
More line length lint cleanup
#247
