+++ R/convert_to_na.R
@@ -48,7 +48,10 @@ convert_to_na <- function(x, ...) {
 convert_to_na.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
     message(insight::format_message(
-      sprintf("Converting values into missing values (`NA`) currently not possible for variables of class `%s`.", class(x)[1])
+      sprintf(
+        "Converting values into missing values (`NA`) currently not possible for variables of class `%s`.",
+        class(x)[1]
+      )
     ))
   }
   x
@@ -180,6 +183,14 @@ convert_to_na.data.frame <- function(x,
     verbose = verbose
   )
 
-  x[select] <- lapply(x[select], convert_to_na, na = na, drop_levels = drop_levels, verbose = verbose, ...)
+  x[select] <- lapply(
+    x[select],
+    convert_to_na,
+    na = na,
+    drop_levels = drop_levels,
+    verbose = verbose,
+    ...
+  )
+
   x
 }
