### 2022-09-04 b93ad2d
Improve message formatting
