### 2022-09-05 aa0b652
data_to_long: improve speed
