### 2022-09-05 b76f8b4
remove string utils
