+++ R/utils_strings.R
@@ -1,19 +1,0 @@
-string_starts_with <- function(pattern, x) {
-  pattern <- paste0("^\\Q", pattern, "\\E")
-  grep(pattern, x, perl = TRUE)
-}
-
-string_contains <- function(pattern, x) {
-  pattern <- paste0("\\Q", pattern, "\\E")
-  grep(pattern, x, perl = TRUE)
-}
-
-string_ends_with <- function(pattern, x) {
-  pattern <- paste0("\\Q", pattern, "\\E$")
-  grep(pattern, x, perl = TRUE)
-}
-
-string_one_of <- function(pattern, x) {
-  m <- unlist(sapply(pattern, grep, x = x, fixed = TRUE, useBytes = TRUE))
-  x[m]
-}
