### 2022-09-05 ea527c8
add tests for demean and data_merge
