+++ tests/testthat/test-demean.R
@@ -14,6 +14,13 @@ test_that("demean works", {
   set.seed(123)
   x <- demean(df, select = c("Sepal.Length", "binary", "Species"), group = "ID")
   expect_snapshot(head(x))
+
+  set.seed(123)
+  expect_equal(
+    demean(df, select = ~ Sepal.Length + binary + Species, group = ~ ID),
+    demean(df, select = c("Sepal.Length", "binary", "Species"), group = "ID")
+  )
+
 })
 
 test_that("demean interaction term", {
@@ -26,4 +33,19 @@ test_that("demean interaction term", {
 
   set.seed(123)
   expect_snapshot(demean(dat, select = c("a", "x*y"), group = "ID"))
+})
+
+test_that("demean shows message if some vars don't exist", {
+  dat <- data.frame(
+    a = c(1, 2, 3, 4, 1, 2, 3, 4),
+    x = c(4, 3, 3, 4, 1, 2, 1, 2),
+    y = c(1, 2, 1, 2, 4, 3, 2, 1),
+    ID = c(1, 2, 3, 1, 2, 3, 1, 2)
+  )
+
+  set.seed(123)
+  expect_message(
+    demean(dat, select = c("foo"), group = "ID"),
+    regexp = "not found"
+  )
 })
