### 2022-09-05 32f7a7d
data_merge: stop if some columns in "by" are not present in both datasets
