### 2022-09-06 ee8c9f3
.select_nse: allow passing arguments through functions and loops
