### 2022-09-06 33de7cc
back to the first implementation of .select_nse
