+++ tests/testthat/test-get_columns.R
@@ -413,3 +413,4 @@ test_that("old solution still works", {
     c("Sepal.Length", "Sepal.Width")
   )
 })
+
