+++ R/select_helpers.R
@@ -1,409 +1,317 @@
 # this function evaluates the select-expression and allows non-standard evaluation
 
 .select_nse <- function(select, data, exclude, ignore_case, regex = FALSE, verbose = FALSE) {
-
-  .check_data(data)
-  cols <- names(data)
-
-  # avoid conflicts for starts_with(), etc.
-  # (has to be in this order, otherwise error because e.g tidyselect is imported
-  # in tidyr)
-  conflicting_packages <- .conflicting_packages(c("poorman", "tidyr", "dplyr", "tidyselect"))
-
-  # Selected vars ---------------------------------------------
-
-  select_deparsed <- insight::safe_deparse(substitute(select, env = parent.frame(1L)))
-  select_is_negated <- FALSE
-  selected <- NULL
-
-  if (!.is_numeric(select_deparsed) &&
-      !.is_object_in_env(select_deparsed) &&
-      .is_negated(select_deparsed)) {
-    select_deparsed <- gsub("^-", "", select_deparsed)
-    select_is_negated <- TRUE
-    tmp <- try(eval(parse(text = select_deparsed)), silent = TRUE)
-    if (!inherits(tmp, "try-error")) {
-      select <- tmp
-    } else if(!.is_select_helper(select_deparsed)) {
-      select <- select_deparsed
-    }
+  # check if data argument is valid
+  if (is.null(data)) {
+    stop(insight::format_message("The `data` argument must be provided."), call. = FALSE)
   }
 
-  if (.is_object_in_env(select_deparsed)) {
-    tmp <- try(get(select_deparsed, envir = parent.frame(2L)), silent = TRUE)
-    if (!inherits(tmp, "try-error")) {
-      select <- tmp
-    }
-  }
-
-  # Deal with range of variable names, e.g var1:var2
-  if (.is_range_character(select_deparsed)) {
-    from_to <- unlist(strsplit(select_deparsed, ":", fixed = TRUE))
-    bound1 <- grep(paste0("^", from_to[1], "$"), cols, ignore.case = ignore_case)
-    bound2 <- grep(paste0("^", from_to[2], "$"), cols, ignore.case = ignore_case)
-    select_deparsed <- cols[bound1:bound2]
-    select <- select_deparsed
-  }
-
-  # Deal with functions with parenthesis, e.g is.numeric()
-  if (length(select_deparsed) == 1 && .is_function(select_deparsed)) {
-    select_deparsed <- gsub("\\(\\)$", "", select_deparsed)
-    select <- get(select_deparsed, envir = parent.frame())
-  }
-
-  # NSE e.g input is Petal.Length
-  if (all(select_deparsed %in% cols)) {
-    selected <- select_deparsed
-    select <- select_deparsed # avoid evaluation below because it would return
-                              # non-evaluable
-  }
-
-  select_is_evaluable <- !inherits(try(eval(select), silent = TRUE), "try-error")
-
-  if (select_is_evaluable) {
-    if (is.null(select)) {
-      selected <- names(data)
-    } else if (is.character(select)) {
-      if (regex) {
-        selected <- cols[which(grepl(select, cols, ignore.case = ignore_case))]
-      } else {
-        selected <- vector(length = length(select))
-        for (i in seq_along(select)) {
-          tmp <- cols[grep(paste0("^", select[i], "$"), cols, ignore.case = ignore_case)]
-          if (length(tmp) > 0) {
-            selected[i] <- tmp
-          }
-        }
-      }
-    } else if (is.numeric(select)) {
-      original_select <- select
-      # negative values in "select" mean "start from the right"
[…]
