+++ tests/testthat/test-coef_var.R
@@ -5,6 +5,5 @@ test_that("coefficient of variation works", {
   expect_equal(coef_var(c(1:10, 100), method = "qcd"), 0.4166667)
   expect_equal(coef_var(mu = 10, sigma = 20), 2)
   expect_equal(coef_var(mu = 10, sigma = 20, method = "unbiased", n = 30), 2.250614)
-  expect_equal(cv(1:10), 0.5504819)
-  expect_equal(distribution_cv(1:10), 0.5504819)
+  expect_equal(distribution_coef_var(1:10), 0.5504819)
 })
