+++ NAMESPACE
@@ -173,7 +173,6 @@ export(compact_character)
 export(compact_list)
 export(convert_na_to)
 export(convert_to_na)
-export(cv)
 export(data_addprefix)
 export(data_addsuffix)
 export(data_adjust)
@@ -205,7 +204,7 @@ export(degroup)
 export(demean)
 export(describe_distribution)
 export(detrend)
-export(distribution_cv)
+export(distribution_coef_var)
 export(distribution_mode)
 export(empty_columns)
 export(empty_rows)
