### 2022-09-06 f59c88a
fix when object is defined in both the function environment and the global environment
