### 2022-09-07 0dedbca
fix when object is in global environment and the whole select helper is passed through a function
