+++ R/data_tabulate.R
@@ -16,8 +16,6 @@
 #' @param ... not used.
 #' @inheritParams find_columns
 #'
-#' @inheritSection center Selection of variables - the `select` argument
-#'
 #' @return A data frame, or a list of data frames, with one frequency table
 #' as data frame per variable.
 #'
