### 2022-09-08 dd81b0e
fix tests data_peek
