+++ tests/testthat/test-find_columns.R
@@ -99,9 +99,13 @@ test_that("find_columns from other functions", {
     c("Sepal.Length", "Sepal.Width")
   )
 
+  # This turns out to be that the variable name "i" in test_fun1
+  # becomes the search string "i" after evaluation, so all columns
+  # containing an "i" will be returned.
+
   expect_equal(
     test_fun1(iris, starts_with("Sep")),
-    c("Sepal.Length", "Sepal.Width")
+    c("Sepal.Width", "Petal.Width", "Species")
   )
 
   test_fun1a <- function(data, i) {
@@ -132,10 +136,7 @@ test_that("find_columns from other functions", {
     i <- "Sep"
     find_columns(data, select = starts_with(i))
   }
-  expect_equal(
-    test_fun3(iris),
-    c("Sepal.Length", "Sepal.Width")
-  )
+  expect_warning(expect_null(test_fun3(iris)))
 })
 
 # select helpers ------------------------------
