+++ R/data_find.R
@@ -18,8 +18,10 @@
 #'     (e.g. `1` or `c(1, 3, 5)`),
 #'   - a vector of negative integers, giving the positions counting from the
 #'     right (e.g., `-1` or `-1:-3`),
-#'   - one of the following select-helpers: `starts_with("")`, `ends_with("")`,
-#'     `contains("")`, a range using `:` or `regex("")`,
+#'   - one of the following select-helpers: `starts_with()`, `ends_with()`,
+#'     `contains()`, a range using `:` or `regex("")`. `starts_with()`,
+#'     `ends_with()`, and  `contains()` accept several patterns, e.g
+#'     `starts_with("Sep", "Petal")`.
 #'   - or a function testing for logical conditions, e.g. `is.numeric()` (or
 #'     `is.numeric`), or any user-defined function that selects the variables
 #'     for which the function returns `TRUE` (like: `foo <- function(x) mean(x) > 3`),
