+++ NEWS.md
@@ -28,9 +28,11 @@ CHANGES
   like the following example. See examples section in the documentation of 
   `data_filter()`.
 
-
 * The `regex` argument was added to functions that use select-helpers and did
   not already have this argument.
+  
+* Select helpers `starts_with()`, `ends_with()`, and  `contains()` now accept
+  several patterns, e.g `starts_with("Sep", "Petal")`.
 
 # datawizard 0.5.1
 
