### 2022-09-09 80a952d
fix with negated custom functions
