### 2022-09-09 13c862d
add some tests when no data, simplify and explain code
