+++ R/select_helpers.R
@@ -1,24 +1,15 @@
 # this function evaluates the select-expression and allows non-standard evaluation
 
 .select_nse <- function(select, data, exclude, ignore_case, regex = FALSE, verbose = FALSE) {
-  # check if data argument is valid
-  if (is.null(data)) {
-    stop(insight::format_message("The `data` argument must be provided."), call. = FALSE)
-  }
 
-  # check data frame input
-  if (!is.null(data) && !is.data.frame(data)) {
-    data <- try(as.data.frame(data), silent = TRUE)
-    if (inherits(data, c("try-error", "simpleError"))) {
-      stop(insight::format_message(
-        "The `data` argument must be a data frame, or an object that can be coerced to a data frame."
-      ), call. = FALSE)
-    }
-  }
+  .check_data(data)
 
+  columns <- colnames(data)
+
+  # directly return all names if select == exclude == NULL
   if (is.null(substitute(select, env = parent.frame())) &
       is.null(substitute(exclude, env = parent.frame()))) {
-    return(names(data))
+    return(columns)
   }
 
   fixed_select <- TRUE
@@ -50,10 +41,10 @@
 
   # select and exclude might also be a (user-defined) function
   if (inherits(select, "function")) {
-    select <- colnames(data)[sapply(data, select)]
+    select <- columns[sapply(data, select)]
   }
   if (inherits(exclude, "function")) {
-    exclude <- colnames(data)[sapply(data, exclude)]
+    exclude <- columns[sapply(data, exclude)]
   }
 
   # if select could not be evaluated (because expression "makes no sense")
@@ -73,10 +64,10 @@
 
   # seems to be no valid column name or index, so try to grep
   if (isFALSE(fixed_select) || isTRUE(regex)) {
-    select <- colnames(data)[grepl(select, colnames(data), ignore.case = ignore_case, perl = TRUE)]
+    select <- columns[grepl(select, columns, ignore.case = ignore_case, perl = TRUE)]
   }
   if (isFALSE(fixed_exclude)) {
-    exclude <- colnames(data)[grepl(exclude, colnames(data), ignore.case = ignore_case, perl = TRUE)]
+    exclude <- columns[grepl(exclude, columns, ignore.case = ignore_case, perl = TRUE)]
   }
   # if exclude = NULL, we want to exclude 0 variables, not all of them
   if (!inherits(exclude, "formula") && length(exclude) == ncol(data)) {
@@ -96,11 +87,7 @@
 
 .evaluate_pattern <- function(x, data = NULL, ignore_case = FALSE, y) {
   fixed <- FALSE
-
-  ## TODO once "data_findcols()" is removed, we can remove the checks
-  # for "is.null(data)" here, because then data is *always* provided and
-  # cannot be NULL. The only place where ".evaluate_pattern()" is called
-  # with no data argument is in "data_findcols()".
+  columns <- colnames(data)
 
   # check if negation is requested
   negate <- !is.null(x) && length(x) == 1 && substr(x, 0, 1) == "-"
@@ -138,8 +125,9 @@
   # However, the expression starts_with("Sep") is part of the error message, so
   # we can use this to retrieve "i".
   suppressWarnings({
-    # if y is in global environment, then evaluation works so we consider that we
-    # need to reevaluate y in this case
+    # if y is in global environment (e.g i = "Petal" in globenv but i = starts_with("Sep")
+    # in the function environment), then evaluation works and we never go through
+    # dynGet() whereas it's needed. So we force y to go in the tryCatch().
[…]
