### 2022-09-12 4dff5ab
.select_nse: allow passing arguments through functions and loops (#255)
* .select_nse: allow passing arguments through functions and loops

* back to the first implementation of .select_nse

* typo [skip ci]

* remove old utils

* add comment to explain the process [skip ci]

* select_nse: deal better with tidyselect error when starts_with is loaded from one of the tidyverse packages

* fix when object is defined in both the function environment and the global environment

* fix when object is in global environment and the whole select helper is passed through a function

* typo [skip ci]

* f
[…]
