+++ R/select_helpers.R
@@ -1,19 +1,15 @@
 # this function evaluates the select-expression and allows non-standard evaluation
 
 .select_nse <- function(select, data, exclude, ignore_case, regex = FALSE, verbose = FALSE) {
-  # check if data argument is valid
-  if (is.null(data)) {
-    stop(insight::format_message("The `data` argument must be provided."), call. = FALSE)
-  }
 
-  # check data frame input
-  if (!is.null(data) && !is.data.frame(data)) {
-    data <- try(as.data.frame(data), silent = TRUE)
-    if (inherits(data, c("try-error", "simpleError"))) {
-      stop(insight::format_message(
-        "The `data` argument must be a data frame, or an object that can be coerced to a data frame."
-      ), call. = FALSE)
-    }
+  .check_data(data)
+
+  columns <- colnames(data)
+
+  # directly return all names if select == exclude == NULL
+  if (is.null(substitute(select, env = parent.frame())) &
+      is.null(substitute(exclude, env = parent.frame()))) {
+    return(columns)
   }
 
   fixed_select <- TRUE
@@ -32,27 +28,37 @@
   }
 
   # check if pattern is a function like "starts_with()"
-  select <- tryCatch(eval(p), error = function(e) NULL)
-  exclude <- tryCatch(eval(p2), error = function(e) NULL)
+  is_p_expr_not_in_globenv <- !is.null(p) && !(insight::safe_deparse(p) %in% ls(globalenv()))
+  is_p2_expr_not_in_globenv <- !is.null(p2) && !insight::safe_deparse(p2) %in% ls(globalenv())
+  if (is_p_expr_not_in_globenv) {
+    select <- tryCatch(eval(p), error = function(e) NULL)
+  } else {
+    select <- NULL
+  }
+  if (is_p2_expr_not_in_globenv) {
+    exclude <- tryCatch(eval(p2), error = function(e) NULL)
+  } else {
+    exclude <- NULL
+  }
 
   # select and exclude might also be a (user-defined) function
   if (inherits(select, "function")) {
-    select <- colnames(data)[sapply(data, select)]
+    select <- columns[sapply(data, select)]
   }
   if (inherits(exclude, "function")) {
-    exclude <- colnames(data)[sapply(data, exclude)]
+    exclude <- columns[sapply(data, exclude)]
   }
 
   # if select could not be evaluated (because expression "makes no sense")
   # try to evaluate and find select-helpers. In this case, set fixed = FALSE,
   # so we can use grepl()
   if (is.null(select)) {
-    evaluated_pattern <- .evaluate_pattern(insight::safe_deparse(p), data, ignore_case = ignore_case)
+    evaluated_pattern <- .evaluate_pattern(insight::safe_deparse(p), data, ignore_case = ignore_case, y = p)
     select <- evaluated_pattern$pattern
     fixed_select <- evaluated_pattern$fixed
   }
   if (is.null(exclude)) {
-    evaluated_pattern <- .evaluate_pattern(insight::safe_deparse(p2), data, ignore_case = ignore_case)
+    evaluated_pattern <- .evaluate_pattern(insight::safe_deparse(p2), data, ignore_case = ignore_case, y = p2)
     exclude <- evaluated_pattern$pattern
     fixed_exclude <- evaluated_pattern$fixed
   }
@@ -60,10 +66,10 @@
 
   # seems to be no valid column name or index, so try to grep
   if (isFALSE(fixed_select) || isTRUE(regex)) {
-    select <- colnames(data)[grepl(select, colnames(data), ignore.case = ignore_case, perl = TRUE)]
+    select <- columns[grepl(select, columns, ignore.case = ignore_case, perl = TRUE)]
   }
   if (isFALSE(fixed_exclude)) {
[…]
