### 2022-09-13 8c1c3d5
selection: using "-" sign with numeric indices removes the columns
