### 2022-09-13 3bd409f
call our own error message if numeric indices both negative and positive
