### 2022-09-13 0bedf78
Selection: using "-" sign with numeric indices removes the columns (#264)
* selection: using "-" sign with numeric indices removes the columns

* call our own error message if numeric indices both negative and positive

* fix tests failing due to behavior change

* bump news and version
