+++ tests/testthat/test-find_columns.R
@@ -10,11 +10,6 @@ test_that("find_columns works as expected", {
   )
 
   expect_equal(
-    find_columns(iris, -starts_with("Sepal", "Petal")),
-    "Species"
-  )
-
-  expect_equal(
     find_columns(iris, ends_with("Width")),
     c("Sepal.Width", "Petal.Width")
   )
@@ -22,11 +17,6 @@ test_that("find_columns works as expected", {
   expect_equal(
     find_columns(iris, ends_with("Length", "Width")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
-  )
-
-  expect_equal(
-    find_columns(iris, -ends_with("Length", "Width")),
-    "Species"
   )
 
   expect_equal(
@@ -47,11 +37,6 @@ test_that("find_columns works as expected", {
   expect_equal(
     find_columns(iris, contains("en", "idt")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
-  )
-
-  expect_equal(
-    find_columns(iris, -contains("en", "idt")),
-    "Species"
   )
 
   expect_equal(
@@ -138,10 +123,62 @@ test_that("find_columns from other functions", {
   )
 })
 
-# select helpers ------------------------------
 test_that("find_columns regex", {
   expect_equal(
     find_columns(mtcars, select = "pg", regex = TRUE),
     find_columns(mtcars, select = "mpg")
+  )
+})
+
+test_that("find_columns works correctly with minus sign", {
+
+  expect_equal(
+    find_columns(iris, -"Sepal.Length"),
+    c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")
+  )
+
+  expect_equal(
+    find_columns(iris, -c("Sepal.Length", "Petal.Width")),
+    c("Sepal.Width", "Petal.Length", "Species")
+  )
+
+  expect_equal(
+    find_columns(iris, -1),
+    c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")
+  )
+
+  expect_error(
+    find_columns(iris, -1:2),
+    regexp = "can't mix negative"
+  )
+
+  expect_equal(
+    find_columns(iris, -(1:2)),
+    c("Petal.Length", "Petal.Width", "Species")
+  )
+
+  expect_equal(
+    find_columns(iris, -c(1, 3)),
+    c("Sepal.Width", "Petal.Width", "Species")
+  )
+
+  expect_equal(
+    find_columns(iris, -starts_with("Sepal", "Petal")),
+    "Species"
+  )
+
+  expect_equal(
+    find_columns(iris, -ends_with("Length", "Width")),
+    "Species"
+  )
+
+  expect_equal(
+    find_columns(iris, -contains("en", "idt")),
+    "Species"
+  )
+
+  expect_equal(
+    find_columns(iris, -c("Sepal.Length", "Petal.Width"), exclude = "Species"),
+    c("Sepal.Width", "Petal.Length")
   )
 })
