+++ tests/testthat/test-get_columns.R
@@ -413,4 +413,3 @@ test_that("old solution still works", {
     c("Sepal.Length", "Sepal.Width")
   )
 })
-
