### 2022-09-14 5b63fe6
add "minus" use with numerics, add programming with selections
