### 2022-09-14 65c9769
add explanation about selecting with formulas
