### 2022-09-14 d11f029
Include more digits in text expected values
