### 2022-09-15 c7a3ba5
example for negating a function in select
