+++ vignettes/selection_syntax.Rmd
@@ -48,7 +48,7 @@ This is the most simple way to select one or several variables. Just use a chara
 vector containing variables names, like in base R.
 
 ```{r}
-get_columns(iris, c("Sepal.Length", "Petal.Width"))
+data_select(iris, c("Sepal.Length", "Petal.Width"))
 ```
 
 ## Unquoted names
@@ -69,7 +69,7 @@ iris %>%
 You can also use formulas to select variables:
 
 ```{r}
-get_columns(iris, ~ Sepal.Length + Petal.Width)
+data_select(iris, ~ Sepal.Length + Petal.Width)
 ```
 
 This was mostly useful to use selection more easily in custom functions before
@@ -83,7 +83,7 @@ In addition to variable names, `select` can also take indices for the variables
 to select in the dataframe.
 
 ```{r}
-get_columns(iris, c(1, 2, 5))
+data_select(iris, c(1, 2, 5))
 ```
 
 
@@ -94,18 +94,17 @@ to all columns and should return `TRUE` or `FALSE`. For example, if we want to
 keep only numeric columns, we can use `is.numeric`.
 
 ```{r}
-get_columns(iris, is.numeric)
+data_select(iris, is.numeric)
 ```
 
-Note that we can provide any custom function to `select`,  
-*provided it returns `TRUE` or `FALSE`* when applied to a column.
+Note that we can provide any custom function to `select`, *provided it returns `TRUE` or `FALSE`* when applied to a column.
 
 ```{r}
 my_function <- function(i) {
   is.numeric(i) && mean(i, na.rm = TRUE) > 3.5
 }
 
-get_columns(iris, my_function)
+data_select(iris, my_function)
 ```
 
 
@@ -118,13 +117,13 @@ To this end, we can use four select helpers: `starts_with()`, `ends_with()`,
 `regex()` takes a single regular expression.
  
 ```{r}
-get_columns(iris, starts_with("Sep", "Peta"))
+data_select(iris, starts_with("Sep", "Peta"))
 
-get_columns(iris, ends_with("dth", "ies"))
+data_select(iris, ends_with("dth", "ies"))
 
-get_columns(iris, contains("pal", "ec"))
+data_select(iris, contains("pal", "ec"))
 
-get_columns(iris, regex("^Sep|ies"))
+data_select(iris, regex("^Sep|ies"))
 ```
 
 ```{r echo=FALSE}
@@ -149,11 +148,11 @@ we can invert our selection.
 The first way is to put a minus sign `"-"` in front of the `select` argument.
 
 ```{r}
-get_columns(iris, -c("Sepal.Length", "Petal.Width"))
+data_select(iris, -c("Sepal.Length", "Petal.Width"))
 
-get_columns(iris, -starts_with("Sep", "Peta"))
+data_select(iris, -starts_with("Sep", "Peta"))
 
-get_columns(iris, -is.numeric)
+data_select(iris, -is.numeric)
 ```
 
 Note that if we use numeric indices, we can't mix negative and positive values. 
@@ -161,18 +160,18 @@ This means that we have to use `select = -(1:2)` if we want to exclude the first
 two columns; `select = -1:2` will *not* work:
 
 ```{r}
-get_columns(iris, -(1:2))
+data_select(iris, -(1:2))
 ```
 
 The second way is to use the argument `exclude`. This argument has the same 
 possibilities as `select`. Although this may not be required in most contexts, 
-if we wanted to, we can use both `select` and `exclude` arguments at the same 
+if we wanted to, we could use both `select` and `exclude` arguments at the same 
 time.
 
 ```{r}
-get_columns(iris, exclude = c("Sepal.Length", "Petal.Width"))
[…]
