### 2022-09-15 137ba1d
move formulas in "useful to know"
