### 2022-09-15 2c45c89
add vignette in NEWS
