### 2022-09-15 aa126a8
Deploying to gh-pages from @ easystats/datawizard@4e41714cd6b93a681b1554b859fd7f8a771b8c99 🚀
