### 2022-09-15 63b4edd
check and explain why x must be ratio scaled
