### 2022-09-15 d9cb700
Update R/descriptives.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
