+++ R/descriptives.R
@@ -94,7 +94,7 @@ coef_var.default <- function(x, verbose = TRUE, ...) {
 #' sandwiches did I eat this week ( = sandwiches * 0.5) and 0 would still mean
 #' "none", and 20 half-sandwiches is still 4 times more than 5 half-sandwiches.
 #'
-#' This means that CV is **NOT** invariance to shifting, but it is to scaling:
+#' This means that CV is **NOT** invariant to shifting, but it is to scaling:
 
 #' ```{r}
 #' sandwiches <- c(0, 4, 15, 0, 0, 5, 2, 7)
