### 2022-09-15 4c60b14
Bump minimum needed R version to `3.6`
https://github.com/easystats/easystats/issues/304
