+++ cran-comments.md
@@ -2,10 +2,11 @@
 
 0 errors | 0 warnings | 0 note
 
-* Fixes failing tests due to `{poorman}` update.
+* Adds new functionality and maintenance updates.
 
 ## revdepcheck results
 
-Expect test failures in a few of the reverse dependencies. The authors of these
-packages have already been informed and they will be submitting revised versions
-soon.
+We checked 15 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
+
+ * We saw 0 new problems
+ * We failed to check 0 packages
