### 2022-09-15 d68f2cc
Add coefficient of variation function (#236)
* Add coefficient of variation function

Related to https://github.com/easystats/datawizard/issues/232, https://github.com/easystats/performance/issues/433

* Remove suggestion that `coef_var()` can be used with models.

Will implement that as `performance_coefvar()` or similar

* code-style, use namespace

* typo in doc

* minor

* Fix logic

* Move mode and CV to descriptives.R

* Bugfix in switch()

* bug fixes and add tests

* Move processing to internal functions for reuse

* Fix documentation

* remove CV alias

* Include more digits in text e
[…]
