### 2022-09-15 #4684 @IndrajeetPatil
> Version: 0.5.0

@osorensen We needed to make a new CRAN release. Therefore, can you please bump the version to `0.6.0`? Thanks.
