+++ NEWS.md
@@ -1,3 +1,5 @@
+# datawizard (development version)
+
 # datawizard 0.6.0
 
 BREAKING CHANGES
