+++ NAMESPACE
@@ -166,7 +166,6 @@ export(centre)
 export(change_code)
 export(change_scale)
 export(coef_var)
-export(coerce_to_numeric)
 export(colnames_to_row)
 export(column_as_rownames)
 export(compact_character)
