+++ DESCRIPTION
@@ -63,7 +63,7 @@ VignetteBuilder:
 Encoding: UTF-8
 Language: en-US
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.1.9000
+RoxygenNote: 7.2.1
 Config/testthat/edition: 3
 Config/Needs/website:
     rstudio/bslib,
