### 2022-09-16 2ac4154
Update detection of tidyselect error message
