### 2022-09-19 4f7edfb
use insight::format_* messages
