+++ R/to_factor.R
@@ -31,9 +31,9 @@ to_factor <- function(x, ...) {
 #' @export
 to_factor.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
-    message(insight::format_message(
+    insight::format_alert(
       sprintf("Converting into factors values currently not possible for variables of class `%s`.", class(x)[1])
-    ))
+    )
   }
   x
 }
