+++ R/data_reverse.R
@@ -45,13 +45,13 @@ reverse_scale <- reverse
 #' @export
 reverse.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
-    message(insight::format_message(
+    insight::format_alert(
       paste0(
         "Variables of class '",
         class(x)[1],
         "' can't be recoded and remain unchanged."
       )
-    ))
+    )
   }
 
   x

[…]
