### 2022-09-19 6980eca
Use `insight::format_*` messages (#269)
use insight::format_* messages
