+++ vignettes/selection_syntax.Rmd
@@ -120,7 +120,7 @@ when we write them.")
 ```
 
 ```{r echo=FALSE}
-row("Note #2: because these functions are not exported, they will not create 
+row("Note #2: because these functions are not exported, they will not create
 conflicts with the ones that come from the `tidyverse` and that have the same name.
 So we can still use `dplyr` and its friends, it won't change anything for selection
 in `datawizard` functions!")
