+++ NAMESPACE
@@ -135,6 +135,8 @@ S3method(to_factor,factor)
 S3method(to_factor,logical)
 S3method(to_factor,numeric)
 S3method(to_numeric,Date)
+S3method(to_numeric,POSIXct)
+S3method(to_numeric,POSIXlt)
 S3method(to_numeric,POSIXt)
 S3method(to_numeric,character)
 S3method(to_numeric,data.frame)
