### 2022-09-25 a7ce980
cran submission 0.6.1
