+++ NEWS.md
@@ -1,3 +1,5 @@
+# datawizard (development version)
+
 # datawizard 0.6.1
 
 * Updates tests for upcoming changes in the `{tidyselect}` package (#267).
