### 2022-09-26 #4684 @osorensen
👋 @tomfaulkenberry, @garretrc, could you please update us on how it's going with your reviews?

Feel free to add comments to your reviews in this thread, or to open issues in the [source repository](https://github.com/easystats/datawizard)
