### 2022-09-26 #4684 @garretrc
I've had the chance to incoporate many of the datawizard functions into my own workflow to test the functionality, I'll wrap up checking some of the remaining functions that I haven't been able to test today.

This package has a lot of useful functions even for experienced R users, so I'm also thinking about ways this package could best communicate that.
