### 2022-09-26 #4684 @garretrc
Review finished. Everything I've tested returns the correct results and runs quickly. A couple comments:

1. I don't think the readme highlights the benefits of datawizard for an experienced R user (yes, a lot of the functions are quick to code up on your own, but you might save 5-10 minutes/a trip to stack overflow by using some datawizard "magic"). However, the easystats ecosystem page addresses this concern a bit, and the datawizard readme seems very detailed already. If I think of a better way to highlight this I'll raise an issue in the repo.

2. When using some of the functions, it's not clear from the tooltip in RStudio what the arguments should be. For example, when I use rescale(), center(), slide(), and others, only the x argument is included in the tooltip. So, when using the package I found myself needing to consult the help documentation in cases where I was expecting to just glance at the tooltip. My understanding is that these functions have different arguments for different data types and different operations, so the second argument will change for different use cases. The flexibility of each function to work with different data types without needing something like dplyr/data.table/base R data.frame manipulation seems more important to the package's mission than the tooltips, but just wanted to share this experience as a user.

The above points are small nitpicks, but I think the selection of functions in the package provides an accessible route to many common data manipulation tasks for new R users. For experienced R users who have previously coded the included operations on their own/through tidyverse or other packages, it seems just as fast to learn a new datawizard function than to look up/code up an equivalent solution. The datawizard functions are well-curated toward tasks that are annoying to remember/implement, so I believe they will be easier to recall and quicker to access than equivalent solutions from more general packages.
