+++ paper/paper.Rmd
@@ -25,7 +25,7 @@ authors:
   
 affiliations:
 - index: 1
-  name: cynkra GmbH, Germany
+  name: cynkra Analytics GmbH, Germany
 - index: 2
   name: Nanyang Technological University, Singapore
 - index: 3
