### 2022-09-27 easystats/datawizard#271 @IndrajeetPatil
Improve `README`
Instead of being a feature dump, it should describe the package goals.
