### 2022-09-27 #4684 @IndrajeetPatil
Dear @garretrc,

Thanks a lot for your wonderful assessment of `{datawizard}`. We are delighted to hear that you found the package useful, and think that it is equally accessible and useful for both naive and advanced R users.

I will respond to your nitpicks point-by-point.

1. You are absolutely right that we could be doing a better job of outlining functionality offered by the package. I have created an issue to remind us (https://github.com/easystats/datawizard/issues/271). Of course, if you think of some concrete suggestions, feel free to either comment on that issue or create a new issue.

2. Unfortunately, this is something that is beyond our control, and is a result of RStudio IDE works.

In this video, I demonstrate how, depending on the supplied argument, if you hit tab, the IDE will provide the accurate argument list. But, if put the cursor on the function name and hit tab, it always displays the same tooltip, irrespective of which S3 method is dispatched.

https://user-images.githubusercontent.com/11330453/192577004-8f343857-d432-4ba3-be2a-37ac8a02a529.mov

Thanks again for your review, and let us know if you have any other comments or suggestions.
