### 2022-09-27 #4684 @garretrc
@IndrajeetPatil I'll leave a comment on that issue if I think of anything!

I didn't know you could supply the first argument like that to change the tooltip, it definitely improves the experience. Not really in your control to make sure a user inputs the first argument before pressing tab.
