### 2022-09-28 ad4e63f
add utils for coercing to dataframe
