+++ R/data_read.R
@@ -15,6 +15,11 @@
 #' @param path_catalog Character string, path to the catalog file. Only relevant
 #' for SAS data files.
 #' @param encoding The character encoding used for the file. Usually not needed.
+#' @param convert_factors Logical, if `TRUE` (default), numeric variables, where
+#' all values have a value label, are assumed to be categorical and converted
+#' into factors. If `FALSE`, no variable types are guessed and no conversion
+#' of numeric variables into factors will be performed. See also section
+#' 'Differences to other packages'.
 #' @param verbose Toggle warnings and messages.
 #' @param ... Arguments passed to the related `read_*()` function.
 #'
@@ -50,10 +55,17 @@
 #' imported value labels will be set as factor levels. Else, if a variable
 #' has _no_ value labels or less value labels than values, the variable is
 #' either converted into numeric or character vector. Value labels are then
-#' preserved as `"labels"` attribute.
+#' preserved as `"labels"` attribute. Use `convert_factors = FALSE` to suppress
+#' automatic detection of factors; numeric variables then will not be converted
+#' to factors.
 #'
 #' @export
-data_read <- function(path, path_catalog = NULL, encoding = NULL, verbose = TRUE, ...) {
+data_read <- function(path,
+                      path_catalog = NULL,
+                      encoding = NULL,
+                      convert_factors = TRUE,
+                      verbose = TRUE,
+                      ...) {
   # extract first valid file from zip-file
   if (.file_ext(path) == "zip") {
     path <- .extract_zip(path)
@@ -66,10 +78,10 @@ data_read <- function(path, path_catalog = NULL, encoding = NULL, verbose = TRUE
     "xls" = ,
     "xlsx" = .read_excel(path, encoding, verbose, ...),
     "sav" = ,
-    "por" = .read_spss(path, encoding, verbose, ...),
-    "dta" = .read_stata(path, encoding, verbose, ...),
-    "sas7bdat" = .read_sas(path, path_catalog, encoding, verbose, ...),
-    .read_unknown(path, verbose, ...)
+    "por" = .read_spss(path, encoding, convert_factors, verbose, ...),
+    "dta" = .read_stata(path, encoding, convert_factors, verbose, ...),
+    "sas7bdat" = .read_sas(path, path_catalog, encoding, convert_factors, verbose, ...),
+    .read_unknown(path, convert_factors, verbose, ...)
   )
 }
 
@@ -96,7 +108,7 @@ data_read <- function(path, path_catalog = NULL, encoding = NULL, verbose = TRUE
     utils::unzip(path, exdir = d)
     path <- file.path(d, dest[1])
   } else {
-    stop("The zip-file does not contain any supported file types.", call. = FALSE)
+    insight::format_error("The zip-file does not contain any supported file types.")
   }
 
   path
@@ -106,40 +118,56 @@ data_read <- function(path, path_catalog = NULL, encoding = NULL, verbose = TRUE
 
 # process imported data from SPSS, SAS or Stata -----------------------
 
-.post_process_imported_data <- function(x, verbose) {
-  if (verbose) {
-    message("Preparing data... Almost there!")
-  }
-  x[] <- lapply(x, function(i) {
-    # save labels
-    value_labels <- attr(i, "labels", exact = TRUE)
-    variable_labels <- attr(i, "label", exact = TRUE)
+.post_process_imported_data <- function(x, convert_factors, verbose) {
+  # user may decide whether we automatically detect variable type or not
+  if (isTRUE(convert_factors)) {
+    if (verbose) {
+      message("Preparing data... Almost there!")
+    }
+    x[] <- lapply(x, function(i) {
+      # only proceed if not all missing
+      if (!all(is.na(i))) {
+        # save labels
[…]
