### 2022-09-28 1552145
Add utils for coercing to dataframe (#272)
* add utils for coercing to dataframe

* fix failing test
