### 2022-09-28 b047c74
allow to skip empty data
