+++ R/data_read.R
@@ -20,6 +20,9 @@
 #' into factors. If `FALSE`, no variable types are guessed and no conversion
 #' of numeric variables into factors will be performed. See also section
 #' 'Differences to other packages'.
+#' @param skip_empty If `TRUE`, "empty" variables (that have completely missing
+#' values, or character vectors that completely have empty character elements)
+#' will be removed from the data frame after import.
 #' @param verbose Toggle warnings and messages.
 #' @param ... Arguments passed to the related `read_*()` function.
 #'
@@ -61,6 +64,7 @@ data_read <- function(path,
                       path_catalog = NULL,
                       encoding = NULL,
                       convert_factors = TRUE,
+                      skip_empty = FALSE,
                       verbose = TRUE,
                       ...) {
   # extract first valid file from zip-file
@@ -69,7 +73,7 @@ data_read <- function(path,
   }
 
   # read data
-  switch(.file_ext(path),
+  out <- switch(.file_ext(path),
     "txt" = ,
     "csv" = .read_text(path, encoding, verbose, ...),
     "xls" = ,
@@ -80,6 +84,26 @@ data_read <- function(path,
     "sas7bdat" = .read_sas(path, path_catalog, encoding, convert_factors, verbose, ...),
     .read_unknown(path, convert_factors, verbose, ...)
   )
+
+  # remove empty variables
+  if (skip_empty) {
+    # find empty variables
+    empty <- sapply(out, function(x) {
+      all(is.na(x)) || (is.character(x) && max(nchar(x), na.rm = TRUE) == 0)
+    })
+    # if any, tell user and remove
+    if (any(empty)) {
+      out <- out[!empty]
+      if (verbose) {
+        insight::format_alert(
+          sprintf("Following %i variables were empty and have been removed:", sum(empty)),
+          text_concatenate(names(empty[empty]))
+        )
+      }
+    }
+  }
+
+  out
 }
 
 
