### 2022-09-28 b0c201a
refactor into small helper function
