### 2022-09-28 ee229e6
fix data_arrange for groups
