+++ NEWS.md
@@ -3,9 +3,6 @@
 * `data_read()` gains a `convert_factors` argument, to turn off automatic
   conversion from numeric variables into factors.
 
-* `data_read()` gains a `skip_empty` argument, skip "empty" (completely missing)
-  variables during import.
-
 # datawizard 0.6.1
 
 * Updates tests for upcoming changes in the `{tidyselect}` package (#267).
