### 2022-09-28 b0e46bc
remove_empty removes empty strings
