+++ R/remove_empty.R
@@ -10,6 +10,8 @@
 #' from the data frame.
 #'
 #' @param x A data frame.
+#' @param include_empty_string If `TRUE` also removes character vectors that
+#' completely have empty elements (i.e. where `nchar() == 0`).
 #'
 #' @return
 #'
@@ -42,11 +44,17 @@
 #' # remove empty columns and rows
 #' remove_empty(tmp)
 #' @export
-empty_columns <- function(x) {
+empty_columns <- function(x, include_empty_string = TRUE) {
   if ((!is.matrix(x) && !is.data.frame(x)) || ncol(x) < 2) {
     vector("numeric")
   } else {
-    which(colSums(is.na(x)) == nrow(x))
+    all_na <- colSums(is.na(x)) == nrow(x)
+    if (include_empty_string) {
+      all_empty <- is.character(x) & max(nchar(x), na.rm = TRUE) == 0
+      which(all_na | all_empty)
+    } else {
+      which(all_na)
+    }
   }
 }
 
@@ -64,9 +72,9 @@ empty_rows <- function(x) {
 
 #' @rdname remove_empty
 #' @export
-remove_empty_columns <- function(x) {
+remove_empty_columns <- function(x, include_empty_string = TRUE) {
   # check if we have any empty columns at all
-  ec <- empty_columns(x)
+  ec <- empty_columns(x, include_empty_string)
 
   # if yes, removing works, else an empty df would be returned
   if (length(ec)) {
