### 2022-09-28 14f170a
remove dplyr < 0.8 code chunks
