+++ R/categorize.R
@@ -317,8 +317,8 @@ categorize.grouped_df <- function(x,
                                   ...) {
   info <- attributes(x)
 
-  # dplyr >= 0.8.0 returns attribute "indices"
-  grps <- attr(x, "groups", exact = TRUE)
+  # works only for dplyr >= 0.8.0
+  grps <- attr(x, "groups", exact = TRUE)[[".rows"]]
 
   # evaluate arguments
   select <- .select_nse(select,
@@ -335,14 +335,6 @@ categorize.grouped_df <- function(x,
   # update processed arguments
   x <- args$x
   select <- args$select
-
-  # dplyr < 0.8.0?
-  if (is.null(grps)) {
-    grps <- attr(x, "indices", exact = TRUE)
-    grps <- lapply(grps, function(x) x + 1)
-  } else {
-    grps <- grps[[".rows"]]
-  }
 
   x <- as.data.frame(x)
   for (rows in grps) {
