+++ R/data_tabulate.R
@@ -162,8 +162,8 @@ data_tabulate.grouped_df <- function(x,
                                      collapse = FALSE,
                                      drop_levels = FALSE,
                                      ...) {
-  # dplyr < 0.8.0 returns attribute "indices"
-  grps <- attr(x, "groups", exact = TRUE)
+  # works only for dplyr >= 0.8.0
+  grps <- attr(x, "groups", exact = TRUE)[[".rows"]]
   group_variables <- NULL
 
   # evaluate arguments
@@ -174,15 +174,6 @@ data_tabulate.grouped_df <- function(x,
     regex = regex,
     verbose = verbose
   )
-
-  # dplyr < 0.8.0?
-  if (is.null(grps)) {
-    grps <- attr(x, "indices", exact = TRUE)
-    grps <- lapply(grps, function(x) x + 1)
-  } else {
-    group_variables <- data_remove(grps, ".rows")
-    grps <- grps[[".rows"]]
-  }
 
   x <- as.data.frame(x)
   out <- list()
