+++ NEWS.md
@@ -1,5 +1,8 @@
 # datawizard (development version)
 
+* `empty_columns()` and `remove_empty_columns()` now also columns that contain
+  only empty characters.
+
 # datawizard 0.6.1
 
 * Updates tests for upcoming changes in the `{tidyselect}` package (#267).
