+++ R/data_tabulate.R
@@ -163,8 +163,9 @@ data_tabulate.grouped_df <- function(x,
                                      drop_levels = FALSE,
                                      ...) {
   # works only for dplyr >= 0.8.0
-  grps <- attr(x, "groups", exact = TRUE)[[".rows"]]
-  group_variables <- NULL
+  grps <- attr(x, "groups", exact = TRUE)
+  group_variables <- data_remove(grps, ".rows")
+  grps <- grps[[".rows"]]
 
   # evaluate arguments
   select <- .select_nse(select,
