### 2022-09-28 23368fe
use vapply
