### 2022-09-28 ebc2608
also remove empty rows with ""
