### 2022-09-29 6189c7a
fix data_arrange for grouped df with NA
