### 2022-09-29 e75fd1d
Update NEWS.md
