### 2022-09-29 d59d8b1
Fix `data_arrange()` for groups (#274)
* fix data_arrange for groups

* fix data_arrange for grouped df with NA

Co-authored-by: Daniel <mail@danielluedecke.de>
