### 2022-09-29 bedb2cf
remove_empty removes empty strings (#275)
* remove_empty removes empty strings

* add test

* update tests

* remove new argument

* use vapply

* docs

* also remove empty rows with ""

* fix test
