+++ R/remove_empty.R
@@ -19,7 +19,12 @@
 #' - For `remove_empty_columns()` and `remove_empty_rows()`, a data frame with
 #' "empty" columns or rows removed, respectively.
 #'
-#' - For `remove_empty`, **both** empty rows and columns will be removed.
+#' - For `remove_empty()`, **both** empty rows and columns will be removed.
+#'
+#' @details For character vectors, empty string values (i.e. `""`) are also
+#' considered as missing value. Thus, if a character vector only contains `NA`
+#' and `""``, it is considered as empty variable and will be removed. Same
+#' applies to observations (rows) that only contain `NA` or `""`.
 #'
 #' @examples
 #' tmp <- data.frame(
@@ -41,12 +46,27 @@
 #'
 #' # remove empty columns and rows
 #' remove_empty(tmp)
+#'
+#' # also remove "empty" character vectors
+#' tmp <- data.frame(
+#'   a = c(1, 2, 3, NA, 5),
+#'   b = c(1, NA, 3, NA, 5),
+#'   c = c("", "", "", "", ""),
+#'   stringsAsFactors = FALSE
+#' )
+#' empty_columns(tmp)
+#'
 #' @export
 empty_columns <- function(x) {
   if ((!is.matrix(x) && !is.data.frame(x)) || ncol(x) < 2) {
     vector("numeric")
   } else {
-    which(colSums(is.na(x)) == nrow(x))
+    all_na <- colSums(is.na(x)) == nrow(x)
+    all_empty <- vapply(x, function(i) {
+      (is.character(i) || is.factor(i)) && max(nchar(as.character(i)), na.rm = TRUE) == 0
+    }, FUN.VALUE = logical(1))
+
+    which(all_na | all_empty)
   }
 }
 
@@ -57,7 +77,7 @@ empty_rows <- function(x) {
   if ((!is.matrix(x) && !is.data.frame(x)) || nrow(x) < 2) {
     vector("numeric")
   } else {
-    which(rowSums(is.na(x)) == ncol(x))
+    which(rowSums(is.na(x) | apply(x, c(1, 2), function(i) nchar(as.character(i)) == 0)) == ncol(x))
   }
 }
 
