### 2022-09-29 276a294
improve performance of empty_rows()
