### 2022-09-29 eb072c2
Improve performance of data_arrange
