+++ R/datawizard-package.R
@@ -1,0 +1,19 @@
+#' \code{datawizard}
+#'
+#' @title datawizard: Easy Data Wrangling and Statistical Transformations
+#'
+#' @description
+#'
+#' A lightweight package to assist in keys steps involved in any data analysis
+#' workflow:
+#'
+#' - wrangling the raw data to get it in the needed form,
+#' - applying preprocessing steps and statistical transformations, and
+#' - compute statistical summaries of data properties and distributions.
+#'
+#' It is also the data wrangling backend for packages in 'easystats' ecosystem.
+#'
+#' @docType package
+#' @aliases datawizard datawizard-package
+#' @name datawizard-package
+"_PACKAGE"
