+++ man/datawizard-package.Rd
@@ -1,0 +1,48 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datawizard-package.R
+\docType{package}
+\name{datawizard-package}
+\alias{datawizard-package}
+\alias{_PACKAGE}
+\alias{datawizard}
+\title{datawizard: Easy Data Wrangling and Statistical Transformations}
+\description{
+A lightweight package to assist in keys steps involved in any data analysis
+workflow:
+\itemize{
+\item wrangling the raw data to get it in the needed form,
+\item applying preprocessing steps and statistical transformations, and
+\item compute statistical summaries of data properties and distributions.
+}
+
+It is also the data wrangling backend for packages in 'easystats' ecosystem.
+}
+\details{
+\code{datawizard}
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://easystats.github.io/datawizard/}
+  \item Report bugs at \url{https://github.com/easystats/datawizard/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets)
+
+Authors:
+\itemize{
+  \item Dominique Makowski \email{dom.makowski@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID}) (@Dom_Makowski)
+  \item Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID}) (@strengejacke)
+  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID})
+  \item Brenton M. Wiernik \email{brenton@wiernik.org} (\href{https://orcid.org/0000-0001-9560-6336}{ORCID}) (@bmwiernik)
+  \item Etienne Bacher \email{etienne.bacher@protonmail.com} (\href{https://orcid.org/0000-0002-9271-5075}{ORCID})
+}
+
+Other contributors:
+\itemize{
+  \item Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID}) (@rempsyc) [contributor]
+}
+
+}
