+++ DESCRIPTION
@@ -19,9 +19,11 @@ Authors@R: c(
            comment = c(ORCID = "0000-0003-4315-6788", Twitter = "@rempsyc"))
   )
 Maintainer: Indrajeet Patil <patilindrajeet.science@gmail.com>
-Description: A lightweight package to easily manipulate, clean, transform,
-    and prepare your data for analysis. It also forms the data wrangling
-    backend for the packages in the 'easystats' ecosystem.
+Description: A lightweight package to assist in keys steps involved in any data 
+    analysis workflow: (1) wrangling the raw data to get it in the needed form, 
+    (2) applying preprocessing steps and statistical transformations, and 
+    (3) compute statistical summaries of data properties and distributions. 
+    It is also the data wrangling backend for packages in 'easystats' ecosystem.
 License: GPL (>= 3)
 URL: https://easystats.github.io/datawizard/
 BugReports: https://github.com/easystats/datawizard/issues
