+++ man/datawizard-package.Rd
@@ -46,3 +46,4 @@ Other contributors:
 }
 
 }
+\keyword{internal}
