### 2022-10-03 4c8b628
fix .has_numeric_rownames()
