### 2022-10-04 7ead5b0
Fix examples for `data_arrange()`
Closes #279
