+++ man/data_arrange.Rd
@@ -23,14 +23,15 @@ A data frame.
 columns.
 }
 \examples{
-\dontrun{
+
 # Arrange using several variables
-data_arrange(head(mtcars), "gear", "carb")
+data_arrange(head(mtcars), c("gear", "carb"))
 
 # Arrange in decreasing order
 data_arrange(head(mtcars), "-carb")
 
+\dontrun{
 # Throw an error if one of the variables specified doesn't exist
-data_arrange(head(mtcars), "gear", "foo", safe = FALSE)
+data_arrange(head(mtcars), c("gear", "foo"), safe = FALSE)
 }
 }
