### 2022-10-04 f3fd531
Bump version to create release for JOSS paper
