### 2022-10-04 easystats/datawizard#280 @osorensen
Editorial comments to JOSS submission
I'll add my editorial comments to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4684) as issues here. This is the first one.

On line 10, please change "keys steps" to "key steps".
