### 2022-10-04 easystats/datawizard#281 @osorensen
Editorial comment to JOSS submission
Here is another editorial comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4684).

On line 21, please add an "a", so the sentence becomes "In effect, {datawizard} provides **a** data processing backend for this entire ecosystem."
