### 2022-10-04 easystats/datawizard#282 @osorensen
Editorial comment to JOSS submission
Here's another editorial comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4684).

On line 24, I suggest removing the ", etc." at the very end. If you have other references in mind than Dowle & Srinivasan (2021) or Eastwood (2021), please mention them explicitly.
