### 2022-10-04 #4684 @IndrajeetPatil
@osorensen Thanks for the checklist.

I have taken all the necessary steps here:

I have created a tagged release for `{datawizard}`. Release version is: [0.6.2](https://github.com/easystats/datawizard/releases/tag/0.6.2).

The DOI for the archived version on Zenodo is: [10.5281/zenodo.7143971](https://zenodo.org/record/7143971#.YzxTDOxBxAc)
