### 2022-10-04 easystats/datawizard#283 @osorensen
Editorial comment to JOSS submission
Here's another editorial comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4684).

On line 26, please use the automatic parenthesized citation rather than adding parentheses manually. You can achieve this by changing from `(@Wickham2019)` to `[@Wickham2019]` in the following line:

https://github.com/easystats/datawizard/blob/0f7f292caf573cdb1e7234e5ed24cea2bcc0f58c/paper/paper.md?plain=1#L59
