### 2022-10-04 easystats/datawizard#284 @osorensen
Editorial comment to JOSS submission
Here's another comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4684).

It's not easy to understand that the table on top of page 4 is the output of the command `describe_distribution(mtcars)`. Could you either add a caption to the table on top of page 4 and reference that table somewhere near line 58, or alternatively just show the R output as you do for the other example commands?
