### 2022-10-04 7f35eae
Address editorial comments to JOSS submission
Closes #280
Closes #281
Closes #282
Closes #283
Closes #284
