+++ R/datawizard-package.R
@@ -4,7 +4,7 @@
 #'
 #' @description
 #'
-#' A lightweight package to assist in keys steps involved in any data analysis
+#' A lightweight package to assist in key steps involved in any data analysis
 #' workflow:
 #'
 #' - wrangling the raw data to get it in the needed form,
