### 2022-10-04 518c6d4
Deploying to gh-pages from @ easystats/datawizard@7f35eae6ba7282a05bb9cd904ac3b6458e13168d 🚀
