### 2022-10-04 37e9bb7
deal with names_sep and names_pattern
