+++ tests/testthat/test-data_reshape.R
@@ -5,7 +5,7 @@ wide_data <- data.frame(replicate(3, sample(1:5)))
 
 test_that("data_reshape works as expected - wide to long", {
   expect_equal(
-    head(data_to_long(wide_data)),
+    head(new_data_to_long(wide_data)),
     data.frame(
       Name = c("X1", "X2", "X3", "X1", "X2", "X3"),
       Value = c(3L, 3L, 2L, 2L, 1L, 3L),
@@ -16,7 +16,7 @@ test_that("data_reshape works as expected - wide to long", {
   )
 
   expect_equal(
-    head(data_to_long(wide_data,
+    head(new_data_to_long(wide_data,
       select = c(1, 2),
       names_to = "Column",
       values_to = "Numbers",
@@ -37,7 +37,7 @@ test_that("data_reshape works as expected - wide to long", {
 
 test_that("data_reshape works as expected - using row names as idvar", {
   data(mtcars)
-  out <- data_to_long(mtcars, select = 2:4)
+  out <- new_data_to_long(mtcars, select = 2:4)
   expect_equal(
     dim(out),
     c(96, 10),
@@ -52,7 +52,7 @@ test_that("data_reshape works as expected - using row names as idvar", {
   )
   expect_equal(
     head(out$value),
-    c(8, 304, 150, 8, 472, 205),
+    c(6, 160, 110, 6, 160, 110),
     ignore_attr = TRUE,
     tolerance = 1e-3
   )
@@ -60,7 +60,7 @@ test_that("data_reshape works as expected - using row names as idvar", {
 
 
 test_that("data_reshape works as expected - long to wide", {
-  long_data <- data_to_long(wide_data, rows_to = "Row_ID")
+  long_data <- new_data_to_long(wide_data, rows_to = "Row_ID")
 
   expect_equal(
     data_to_wide(
@@ -93,7 +93,7 @@ test_that("data_reshape works as expected - long to wide", {
   )
 
   # colnames
-  long_data <- data_to_long(wide_data, select = c("X2", "X3"))
+  long_data <- new_data_to_long(wide_data, select = c("X2", "X3"))
   wide <- data_to_wide(long_data, names_from = "name", values_from = "value")
   expect_equal(colnames(wide), colnames(wide_data))
 })
@@ -104,7 +104,7 @@ test_that("data_reshape works as expected - complex dataset", {
 
   data <- psych::bfi
 
-  long <- data_to_long(data,
+  long <- new_data_to_long(data,
     select = regex("\\d"),
     names_to = "Item",
     values_to = "Score",
@@ -125,14 +125,14 @@ test_that("data_reshape works as expected - complex dataset", {
   expect_snapshot(str(wide))
 
 
-  long1 <- data_to_long(data,
+  long1 <- new_data_to_long(data,
     select = starts_with("A"),
     names_to = "Item",
     values_to = "Score",
     rows_to = "Participant"
   )
 
-  long2 <- data_to_long(data,
+  long2 <- new_data_to_long(data,
     select = c("A1", "A2", "A3", "A4", "A5"),
     names_to = "Item",
     values_to = "Score",
@@ -150,14 +150,14 @@ test_that("data_reshape works as expected - complex dataset", {
   expect_equal(nrow(long1), nrow(long2))
 
 
-  long1 <- data_to_long(data,
+  long1 <- new_data_to_long(data,
     select = starts_with("a"),
     names_to = "Item",
     values_to = "Score",
     rows_to = "Participant"
   )
 
-  long2 <- data_to_long(data,
+  long2 <- new_data_to_long(data,
     select = "age",
     names_to = "Item",
     values_to = "Score",
[…]
