### 2022-10-04 ffb5e3e
remove unused helper
