+++ R/data_to_wide_new.R
@@ -1,16 +1,16 @@
 new_data_to_wide <- function(data,
-                         id_cols = NULL,
-                         values_from = "Value",
-                         names_from = "Name",
-                         names_sep = "_",
-                         names_prefix = "",
-                         names_glue = NULL,
-                         values_fill = NULL,
-                         verbose = TRUE,
-                         ...,
-                         colnames_from,
-                         rows_from,
-                         sep) {
+                             id_cols = NULL,
+                             values_from = "Value",
+                             names_from = "Name",
+                             names_sep = "_",
+                             names_prefix = "",
+                             names_glue = NULL,
+                             values_fill = NULL,
+                             verbose = TRUE,
+                             ...,
+                             colnames_from,
+                             rows_from,
+                             sep) {
 
   if (!missing(colnames_from)) {
     .is_deprecated("colnames_from", "names_from")
@@ -30,7 +30,7 @@ new_data_to_wide <- function(data,
       names_sep <- sep
     }
   }
-  old_names <- names(data)
+  current_colnames <- names(data)
 
   # Preserve attributes
   if (inherits(data, "tbl_df")) {
@@ -39,26 +39,94 @@ new_data_to_wide <- function(data,
   } else {
     tbl_input <- FALSE
   }
+
   variable_attr <- lapply(data, attributes)
 
   not_selected <- setdiff(names(data), c(names_from, values_from))
   not_unstacked <- data[, not_selected, drop = FALSE]
   not_unstacked <- unique(not_unstacked)
 
-  unstack_fml <- as.formula(
-    paste(
-      paste(values_from, collapse = " + "),
-      "~",
-      paste(names_from, collapse = " + ")
+  # unstack doesn't create NAs for combinations that don't exist (contrary to
+  # reshape), so it makes a list of unequal length and not a dataframe.
+  # Solution: complete the dataset before unstacking, but it takes some time.
+
+  foo <- as.data.frame(table(data[, names_from]))
+  if (insight::n_unique(foo$Freq) == 1 ||
+      (insight::n_unique(foo$Freq) == 2 && 0 %in% unique(foo$Freq))) {
+    new_data <- data
+  } else {
+    new_data <- unique(expand.grid(data[, c(not_selected, names_from)]))
+    new_data <- data_merge(new_data, data, by = c(not_selected, names_from))
+  }
+
+  # Fill missing values (before converting to wide)
+  if (!is.null(values_fill)) {
+    if (length(values_fill) == 1) {
+      if (is.numeric(new_data[[values_from]])) {
+        if (!is.numeric(values_fill)) {
+          insight::format_error(paste0("`values_fill` must be of type numeric."))
+        } else {
+          new_data <- convert_na_to(new_data, replace_num = values_fill)
+        }
[…]
