+++ tests/testthat/test-data_reshape.R
@@ -63,7 +63,7 @@ test_that("data_reshape works as expected - long to wide", {
   long_data <- new_data_to_long(wide_data, rows_to = "Row_ID")
 
   expect_equal(
-    data_to_wide(
+    new_data_to_wide(
       long_data,
       names_from = "name",
       values_from = "value",
@@ -83,7 +83,7 @@ test_that("data_reshape works as expected - long to wide", {
   # warning for column name collision
   long_data$X1 <- 5
   expect_error(
-    data_to_wide(
+    new_data_to_wide(
       long_data,
       names_from = "name",
       values_from = "value",
@@ -94,7 +94,7 @@ test_that("data_reshape works as expected - long to wide", {
 
   # colnames
   long_data <- new_data_to_long(wide_data, select = c("X2", "X3"))
-  wide <- data_to_wide(long_data, names_from = "name", values_from = "value")
+  wide <- new_data_to_wide(long_data, names_from = "name", values_from = "value")
   expect_equal(colnames(wide), colnames(wide_data))
 })
 
@@ -117,7 +117,7 @@ test_that("data_reshape works as expected - complex dataset", {
   long$Item <- gsub("[A-Z]", "", long$Item)
   long$Item <- paste0("I", long$Item)
 
-  wide <- data_to_wide(long,
+  wide <- new_data_to_wide(long,
     names_from = "Item",
     values_from = "Score"
   )
@@ -312,9 +312,9 @@ test_that("data_reshape works as expected - select-helper inside functions, usin
 })
 
 
-test_that("reshape_wider, names_prefix works", {
+test_that("new_data_to_wide, names_prefix works", {
   out <- fish_encounters %>%
-    reshape_wider(
+    new_data_to_wide(
       names_from = "station",
       values_from = "seen",
       names_prefix = "foo_"
@@ -329,11 +329,11 @@ test_that("reshape_wider, names_prefix works", {
   )
 })
 
-test_that("reshape_wider, values_fill errors when wrong type", {
+test_that("new_data_to_wide, values_fill errors when wrong type", {
   ### Should be numeric
   expect_error(
     fish_encounters %>%
-      reshape_wider(
+      new_data_to_wide(
         names_from = "station",
         values_from = "seen",
         values_fill = "a"
@@ -342,7 +342,7 @@ test_that("reshape_wider, values_fill errors when wrong type", {
   )
   expect_error(
     fish_encounters %>%
-      reshape_wider(
+      new_data_to_wide(
         names_from = "station",
         values_from = "seen",
         values_fill = factor("a")
@@ -364,7 +364,7 @@ test_that("reshape_wider, values_fill errors when wrong type", {
 
   expect_error(
     contacts %>%
-      reshape_wider(
+      new_data_to_wide(
         names_from = "field",
         values_from = "value",
         values_fill = 1
@@ -373,7 +373,7 @@ test_that("reshape_wider, values_fill errors when wrong type", {
   )
   expect_error(
     contacts %>%
-      reshape_wider(
+      new_data_to_wide(
         names_from = "field",
         values_from = "value",
         values_fill = factor("a")
@@ -385,7 +385,7 @@ test_that("reshape_wider, values_fill errors when wrong type", {
   contacts$value <- as.factor(contacts$value)
   expect_error(
     contacts %>%
-      reshape_wider(
+      new_data_to_wide(
         names_from = "field",
[…]
