### 2022-10-06 dd92c9a
polish code for data_to_wide
