+++ R/data_to_wide_new.R
@@ -49,47 +49,59 @@ new_data_to_wide <- function(
   not_unstacked <- unique(not_unstacked)
 
   # unstack doesn't create NAs for combinations that don't exist (contrary to
-  # reshape), so it makes a list of unequal length and not a dataframe.
-  # Solution: complete the dataset before unstacking, but it takes some time.
+  # reshape), so we need to complete the dataset before unstacking.
 
   new_data <- data
+
+  # create an id with all variables that are not in names_from or values_from
+  # so that we can create missing combinations between this id and names_from
   if (length(not_selected) > 1) {
-    new_data$abcdefgh <- do.call(paste, c(new_data[, not_selected, drop = FALSE], sep = "_"))
+    new_data$temporary_id <- do.call(paste, c(new_data[, not_selected, drop = FALSE], sep = "_"))
   } else if (length(not_selected) == 1) {
-    new_data$abcdefgh <- new_data[[not_selected]]
+    new_data$temporary_id <- new_data[[not_selected]]
   } else {
-    new_data$abcdefgh <- seq_len(nrow(new_data))
+    new_data$temporary_id <- seq_len(nrow(new_data))
   }
 
   # check that all_groups have all possible values for names_from
   # If not, need to complete the dataset with NA for values_from where names_from
   # didn't exist
-  n_row_per_group <- table(new_data$abcdefgh)
+  n_rows_per_group <- table(new_data$temporary_id)
+  n_values_per_group <- insight::n_unique(n_rows_per_group)
 
-  not_all_cols_selected <- length(not_selected) > 0
+  not_all_cols_are_selected <- length(not_selected) > 0
 
-  incomplete_groups <- (insight::n_unique(n_row_per_group) > 1 &&
-    !all(unique(n_row_per_group) %in% insight::n_unique(new_data[, names_from]))) ||
-    (insight::n_unique(n_row_per_group) == 1 &&
-    unique(n_row_per_group) < length(unique(new_data[, names_from])))
+  incomplete_groups <-
+    (n_values_per_group > 1 &&
+       !all(unique(n_rows_per_group) %in% insight::n_unique(new_data[, names_from]))
+     ) ||
+    (n_values_per_group == 1 &&
+       unique(n_rows_per_group) < length(unique(new_data[, names_from]))
+    )
 
-  if (not_all_cols_selected && incomplete_groups) {
+  # create missing combinations
 
-    expanded <- expand.grid(unique(new_data[["abcdefgh"]]), unique(new_data[[names_from]]))
-    names(expanded) <- c("abcdefgh", names_from)
-    new_data <- data_merge(new_data, expanded, join = "full", by = c("abcdefgh", names_from), sort = FALSE)
+  if (not_all_cols_are_selected && incomplete_groups) {
 
+    expanded <- expand.grid(unique(new_data[["temporary_id"]]), unique(new_data[[names_from]]))
+    names(expanded) <- c("temporary_id", names_from)
+    new_data <- data_merge(new_data, expanded,
+                           join = "full", by = c("temporary_id", names_from),
+                           sort = FALSE)
+
+    # creation of missing combinations was done with a temporary id, so need
+    # to fill columns that are not selected in names_from or values_from
     new_data[, not_selected] <- lapply(not_selected, function(x) {
-      ave(new_data[[x]], new_data$abcdefgh, FUN = function(y) {
+      ave(new_data[[x]], new_data$temporary_id, FUN = function(y) {
[…]
