### 2022-10-07 3d107ff
polish code for data_to_long
