+++ NAMESPACE
@@ -163,6 +163,7 @@ S3method(winsorize,data.frame)
 S3method(winsorize,factor)
 S3method(winsorize,logical)
 S3method(winsorize,numeric)
+export()
 export(adjust)
 export(categorize)
 export(center)

[…]
