+++ NAMESPACE
@@ -163,7 +163,6 @@ S3method(winsorize,data.frame)
 S3method(winsorize,factor)
 S3method(winsorize,logical)
 S3method(winsorize,numeric)
-export()
 export(adjust)
 export(categorize)
 export(center)
@@ -221,6 +220,8 @@ export(kurtosis)
 export(normalize)
 export(object_has_names)
 export(object_has_rownames)
+export(old_data_to_long)
+export(old_data_to_wide)
 export(print_html)
 export(print_md)
 export(ranktransform)
