### 2022-10-07 8b9630e
Update R/data_to_long.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
