+++ NAMESPACE
@@ -220,8 +220,6 @@ export(kurtosis)
 export(normalize)
 export(object_has_names)
 export(object_has_rownames)
-export(old_data_to_long)
-export(old_data_to_wide)
 export(print_html)
 export(print_md)
 export(ranktransform)
