### 2022-10-08 1629711
more informative error (https://github.com/easystats/parameters/issues/788)
