+++ R/data_relocate.R
@@ -96,14 +96,14 @@ data_relocate <- function(data,
   # Find new positions
   if (!is.null(before)) {
     before <- before[before %in% data_cols][1] # Take first that exists (if vector is supplied)
-    if (length(before) != 1) {
+    if (length(before) != 1 || is.na(before)) {
       stop("The column passed to `before` wasn't found. Possibly mispelled.", call. = FALSE)
     }
     where <- min(match(before, data_cols))
     position <- c(setdiff(position, where), where)
   } else if (!is.null(after)) {
     after <- after[after %in% data_cols][1] # Take first that exists (if vector is supplied)
-    if (length(after) != 1) {
+    if (length(after) != 1 || is.na(after)) {
       stop("The column passed to `after` wasn't found. Possibly mispelled.", call. = FALSE)
     }
     where <- max(match(after, data_cols))
