+++ R/data_relocate.R
@@ -97,14 +97,14 @@ data_relocate <- function(data,
   if (!is.null(before)) {
     before <- before[before %in% data_cols][1] # Take first that exists (if vector is supplied)
     if (length(before) != 1 || is.na(before)) {
-      stop("The column passed to `before` wasn't found. Possibly mispelled.", call. = FALSE)
+      insight::format_error("The column passed to `before` wasn't found. Possibly mispelled.")
     }
     where <- min(match(before, data_cols))
     position <- c(setdiff(position, where), where)
   } else if (!is.null(after)) {
     after <- after[after %in% data_cols][1] # Take first that exists (if vector is supplied)
     if (length(after) != 1 || is.na(after)) {
-      stop("The column passed to `after` wasn't found. Possibly mispelled.", call. = FALSE)
+      insight::format_error("The column passed to `after` wasn't found. Possibly mispelled.")
     }
     where <- max(match(after, data_cols))
     position <- c(where, setdiff(position, where))
