### 2022-10-08 ecc7522
Deploying to gh-pages from @ easystats/datawizard@4b995b5b237e6eea57fbd2b29b72ca30a3b69d09 🚀
