+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   selection_syntax: selection_syntax.html
   standardize_data: standardize_data.html
   tidyverse_translation: tidyverse_translation.html
-last_built: 2022-10-08T22:43Z
+last_built: 2022-10-09T08:30Z
 urls:
   reference: https://easystats.github.io/datawizard/reference
   article: https://easystats.github.io/datawizard/articles
