+++ NEWS.md
@@ -1,3 +1,10 @@
+# datawizard 0.6.2.1
+
+MAJOR CHANGES
+
+* There is new a publication about the `{datawizard}` package: 
+  Patil et al. (2022) <doi:10.21105/joss.04684>.
+
 # datawizard 0.6.2
 
 BREAKING CHANGES
@@ -10,13 +17,15 @@ BREAKING CHANGES
   `remove_empty_rows()` remove observations that completely have missing or
   empty character values.
 
-CHANGES
-
-* `data_arrange()` now works with data frames that were grouped using
-  `data_group()` (#274).
+MINOR CHANGES
 
 * `data_read()` gains a `convert_factors` argument, to turn off automatic
   conversion from numeric variables into factors.
+
+BUG FIXES
+
+* `data_arrange()` now works with data frames that were grouped using
+  `data_group()` (#274).
 
 # datawizard 0.6.1
 
