### 2022-10-09 86abfe9
Deploying to gh-pages from @ easystats/datawizard@23b6d4e75cb7815d9a07a747b63aa8ed785558ec 🚀
