+++ vignettes/selection_syntax.Rmd
@@ -40,6 +40,10 @@ row <- function(...) {
 }
 ```
 
+This vignette can be referred to by citing the following:
+
+Patil et al., (2022). datawizard: An R Package for Easy Data Preparation and Statistical Transformations. *Journal of Open Source Software*, *7*(78), 4684, https://doi.org/10.21105/joss.04684
+
 # Selecting variables
 
 ## Quoted names
