### 2024-04-12 #4684 @xuanxu
The error is caused by the name of the directory containing the paper having a space in it : `/JOSS files/`.
Please @IndrajeetPatil can you rename that folder?
